% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-huewarp.R
\name{add_colorjam_preset}
\alias{add_colorjam_preset}
\title{Add colorjam hue warp preset}
\usage{
add_colorjam_preset(preset, h1, h2, verbose = TRUE, ...)
}
\arguments{
\item{preset}{\code{character} string with the preset name.}

\item{h1, h2}{\code{numeric} vectors of equal length, or \code{NULL} to
remove an existing preset.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Add colorjam hue warp preset
}
\examples{
h1 <- c(8, 30, 65,
   120, 200, 240,
   260, 280, 330)
h2 <- c(115, 35.9, -3.7,
   -3.7, -3.7, -3.7,
   -70.9, -189.6, -229.2)
add_colorjam_preset("dichromat2", h1=h1, h2=h2)
color_pie(rainbowJam(n=10,
   preset="dichromat2",
   h1=h1, h2=h2,
   phase=c(2,1,3,4,5,6)))

add_colorjam_preset("dichromat2", h1=NULL, h2=NULL)
colorjam_presets()

}
\seealso{
Other colorjam hue warp: 
\code{\link{add_colorjam_step}()},
\code{\link{adjust_hue_warp}()},
\code{\link{approx_degrees}()},
\code{\link{colorjam_presets}()},
\code{\link{colorjam_steps}()},
\code{\link{display_degrees}()},
\code{\link{h2hwOptions}()},
\code{\link{h2hw}()},
\code{\link{hw2h}()},
\code{\link{mean_angle}()}
}
\concept{colorjam hue warp}
