% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-steps.R
\name{add_colorjam_step}
\alias{add_colorjam_step}
\title{Add colorjam chroma/luminance step}
\usage{
add_colorjam_step(step, step_list, verbose = TRUE, ...)
}
\arguments{
\item{step}{\code{character} string with the step name.}

\item{step_list}{\code{list} with elements \code{"C"} and \code{"L"} which each contain
a \code{numeric} vector with values in the range \code{c(0, 100)}.
In future the \code{step_list} may permit other color channels, such
as \code{"S"} and \code{"L"} for use with HSL color space, or \code{"S"} and \code{"V"}
for use with HSV color space.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Add colorjam chroma/luminance steps to adjust a series of color hues
to visibly distinct categorical colors.
}
\examples{
colorjam_steps()
colorjam_steps("v24")

Cvals <- c(110, 88, 99,  80, 105, 97)
Lvals <- c(51, 82, 70, 60, 77, 65)
add_colorjam_step("new_v24", step_list=list(C=Cvals, L=Lvals))
colorjam_steps()
colorjam_steps("new_v24")

color_pie(rainbowJam(n=10,
   step="new_v24",
   phase=c(2,1,3,4,5,6)))

add_colorjam_step("new_v24", step_list=NULL)
colorjam_steps()

}
\seealso{
Other colorjam hue warp: 
\code{\link{add_colorjam_preset}()},
\code{\link{adjust_hue_warp}()},
\code{\link{approx_degrees}()},
\code{\link{colorjam_presets}()},
\code{\link{colorjam_steps}()},
\code{\link{display_degrees}()},
\code{\link{h2hwOptions}()},
\code{\link{h2hw}()},
\code{\link{hw2h}()},
\code{\link{mean_angle}()}
}
\concept{colorjam hue warp}
