% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-huewarp.R
\name{adjust_hue_warp}
\alias{adjust_hue_warp}
\title{Adjust the color hue warp effect}
\usage{
adjust_hue_warp(
  h1 = NULL,
  h2 = NULL,
  direction = NULL,
  preset = getOption("colorjam.preset", "custom"),
  h1_shift = 0,
  h2_shift = 0,
  reverse_h2 = FALSE,
  ...
)
}
\arguments{
\item{h1, h2}{\code{numeric} or \code{NULL}}

\item{preset}{\code{character} string used to define \code{h1} and \code{h2} when those
values are not defined specifically.}

\item{h1_shift}{\code{numeric} angle in degrees to shift the \code{h1} hue.
It is recommended to shift \code{h2} and not \code{h1}.}

\item{h2_shift}{\code{numeric} angle in degrees to shift the \code{h2} hue.
It is recommended to shift \code{h2} and not \code{h1}.}

\item{reverse_h2}{\code{logical} indicating whether to reverse the order
of values in \code{h2}.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} of color warp angles with elements \code{"h1"} and \code{"h2"},
suitable for use by \code{h2hw()} and \code{hw2h()}.
}
\description{
Adjust the color hue warp effect, experimental
}
\details{
This function is currently being tested as an approach to adjust
the position and order of the warp color hues. For example,
the initial use case is to "rotate" the color wheel so the starting
color is not always red. Also, the color wheel can be reversed
so the color sequence is reversed.
}
\examples{
new_h1h2 <- adjust_hue_warp(preset="dichromat", h2_shift=0, reverse_h2=FALSE)
add_colorjam_preset("temp", h1=new_h1h2$h1, h2=new_h1h2$h2)
rj_0 <- rainbowJam(n=24, preset="temp", step="v23")
color_pie(rj_0, radius=1,
   main="dichromat color wheel\nstep='v23'")

rj_0 <- rainbowJam(n=24, preset="temp", step="v23", phase=c(1,2,4,3,6,5))
color_pie(rj_0, radius=1,
   main="dichromat color wheel\nstep='v23'\ncustom phase")

n <- 24;
new_h1h2 <- adjust_hue_warp(preset="dichromat", h2_shift=-120, reverse_h2=FALSE)
add_colorjam_preset("temp", h1=new_h1h2$h1, h2=new_h1h2$h2, direction=1)
rj_120 <- rainbowJam(n=n, preset="temp", step="v23",
   nameStyle="n")
color_pie(rj_120, radius=1,
   main="dichromat color wheel rotated -120 degrees\nstep='v23'")

new_h1h2 <- adjust_hue_warp(preset="dichromat", h2_shift=0, reverse_h2=TRUE)
add_colorjam_preset("temp", h1=new_h1h2$h1, h2=new_h1h2$h2,
   direction=new_h1h2$direction)
rj_0rev <- rainbowJam(n=n, preset="temp", step="v23")
names(rj_0rev) <- seq_len(n)
color_pie(rj_0rev, radius=1, main="dichromat color wheel (reversed)")

new_h1h2 <- adjust_hue_warp(preset="dichromat", h2_shift=90, reverse_h2=FALSE)
add_colorjam_preset("temp", h1=new_h1h2$h1, h2=new_h1h2$h2,
   direction=new_h1h2$direction)
rj_90 <- rainbowJam(n=n, preset="temp", step='v23')
color_pie(rj_90, radius=1,
   main="color wheel rotated 90 degrees\nstep='v23'")

# RGB rotated to start at yellow, then red, then blue
new_h1h2 <- adjust_hue_warp(preset="rgb", h2_shift=-70, reverse_h2=TRUE)
add_colorjam_preset("temp", h1=new_h1h2$h1, h2=new_h1h2$h2,
   direction=new_h1h2$direction)
n <- 10
rgb_rev <- rainbowJam(n=n,
   preset="temp", step='v24')
color_pie(rgb_rev,
   main="RGB color wheel rotated -30 degrees (reversed)\nstep='v24'")

# same as above except using ryb3
ryb_h1h2 <- adjust_hue_warp(preset="ryb", h2_shift=-110, reverse_h2=TRUE)
add_colorjam_preset("temp", h1=ryb_h1h2$h1, h2=ryb_h1h2$h2,
   direction=new_h1h2$direction)
n <- 10
ryb_rev <- rainbowJam(n=n,
   #phase=c(1,4,5,2,6,3),
   preset="temp", step='v24')
color_pie(ryb_rev,
   main="RYB color wheel rotated -110 degrees (reversed)\nstep='v24'")

# remove "temp" preset
add_colorjam_preset(preset="temp", h1=NULL)

}
\seealso{
Other colorjam hue warp: 
\code{\link{add_colorjam_preset}()},
\code{\link{add_colorjam_step}()},
\code{\link{approx_degrees}()},
\code{\link{colorjam_presets}()},
\code{\link{colorjam_steps}()},
\code{\link{display_degrees}()},
\code{\link{h2hw}()},
\code{\link{h2hwOptions}()},
\code{\link{hcl_to_hsl_hue}()},
\code{\link{hsl_to_hcl_hue}()},
\code{\link{hw2h}()},
\code{\link{mean_angle}()},
\code{\link{plot_colorjam_preset}()},
\code{\link{plot_colorjam_steps}()},
\code{\link{remap_colorjam_preset}()},
\code{\link{validate_colorjam_preset}()}
}
\concept{colorjam hue warp}
