% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-color.R
\name{closest_named_color}
\alias{closest_named_color}
\title{Closest colorjam named_colors}
\usage{
closest_named_color(
  x,
  colorSet = named_colors,
  C_min = Cgrey,
  Cgrey = getOption("jam.Cgrey", 5),
  showPalette = FALSE,
  colorModel = c("hcl", "LUV"),
  Hwt = 2.5,
  Cwt = 1,
  Lwt = 4,
  warpHue = TRUE,
  preset = "ryb",
  method = "maximum",
  returnType = c("color", "name", "match"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{character vector of colors, either in hex format or any
valid color in R.}

\item{colorSet}{\code{character} vector of colors, by default \code{named_colors}
with provides 4,447 total hex colors, each with human-assigned
color name. These colors also include hex colors from R \code{colors()}
which were not already included in the reference colors.}

\item{C_min, Cgrey}{\code{numeric} minimum color Chroma filter applied to
handle greyscale colors. In most practical cases \code{C_min} and \code{Cgrey}
should be the same value. Note \code{Cgrey} is used as an option in
\code{jamba::make_styles()}, \code{jamba::applyCLrange()} for similar
use cases, so it is used here as well: \code{getOption("jamba.Cgrey", 5)}.
\itemize{
\item \code{C_min} is applied to \code{colorSet} to require the closest matching
color to have at least this color Chroma (saturation).
\item \code{Cgrey} is applied to \code{x} to determine if the input color itself
is considered greyscale, in which case it should not be matched
with saturated colors since there is no reliable color hue.
Instead, the subset of \code{colorSet} with Chroma below \code{C_min} is used
for color-matching.
}

Reworded in short:
\itemize{
\item Colors in \code{x} with Chroma above \code{Cgrey} are matched with
colors in \code{colorSet} with Chroma above \code{C_min}.
\item Colors in \code{x} with Chroma below \code{Cgrey} are matched with
colors in \code{colorSet} with Chroma below \code{C_min}.
\item The end result should be that saturated input colors match
saturated reference colors, and unsaturated input colors match
unsaturated reference colors.
}}

\item{showPalette}{\code{logical} indicating whether to display the input
colors and resulting closest matching colors by using
\code{jamba::showColors()}.}

\item{colorModel}{\code{character} color model to use:
\itemize{
\item \code{"hcl"}: default, uses HCL provided by \code{jamba::col2hcl()} which
uses the equivalent of \code{colorspace::polarLUV()} and considers
color hues in terms of 360 degree angles along a color wheel.
\item \code{"LUV"}: uses CIELUV color space, provided by \code{colorspace::LUV()}
which encodes the angular color hue in 3-D Cartesian space,
allowing comparisons using Euclidean distance.
}}

\item{Hwt, Cwt, Lwt}{\code{numeric} relative weights for each dimension of
HCL colors, for the H, C, and L channels, respectively.}

\item{warpHue}{\code{logical} indicating whether to perform the hue warp
operation using \code{h2hw()} which improves the ability to match
colors between orange and green.}

\item{preset}{\code{character} string to define the color wheel used
when matching input colors \code{x} to colors in \code{colorSet}.
This preset is used with \code{h2hw()} and \code{hw2h()}.
The default \code{preset="ryb"} allows greatest distinction in colors
without imposing additional restrictions such as by \code{preset="dichromat"}
which would only match color-safe colors. The purpose here is
to identify and label colors based upon a reference set of colors.}

\item{method}{\code{character} string passed to \code{stats::dist()}. The default
\code{method="maximum"} works well for \code{colorModel="hcl"}, and
assigns distance using the largest distance across
the three color coordinates H, C, and L. It requires the best
overall match across all three coordinates rather than any weighted
combination of coordinate distances. Other methods in testing allowed
matches of different color hues when luminance and chroma values
were very similar.
With  \code{colorModel="LUV"} we recommend using \code{method="euclidean"},
which seems to work well with projected color coordinates
L, U, and V. The U, and V coordinates are roughly the angular
color hue projected into a flat plane, the L describing Luminance.}

\item{returnType}{\code{character} type of data to return:
\itemize{
\item \code{"color"} returns the color values in \code{colorSet}, which by default
are color names from \code{grDevices::colors()}
\item \code{"name"} returns \code{names(colorSet)} if they exist, otherwise
values from \code{colorSet}
\item \code{"match"} returns an integer vector as an index to \code{colorSet}
}}

\item{verbose}{\code{logical} whether to print verbose output.}
}
\description{
Closest colorjam named_colors for a vector of colors
}
\seealso{
Other colorjam core: 
\code{\link{blend_colors}()},
\code{\link{closestRcolor}()},
\code{\link{color_complement}()},
\code{\link{colors_to_df}()},
\code{\link{group2colors}()},
\code{\link{rainbowJam}()},
\code{\link{sort_colors}()},
\code{\link{subset_colors}()}
}
\concept{colorjam core}
