% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-gradient.R
\name{col_linear_xf}
\alias{col_linear_xf}
\title{Linear color interpolation function with adjustable range, baseline, and floor}
\usage{
col_linear_xf(
  x = 1,
  floor = 0,
  baseline = 0,
  lens = 0,
  n = 6,
  colramp = "Purples",
  open_floor = FALSE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{numeric} value used as a threshold, where numeric
values at or above this value \code{x} are assigned the last
color in the color gradient. Negative values at or below
this negative value \code{-x} are assigned the first color
in the color gradient.}

\item{floor}{\code{numeric} optional value where numeric values
between \code{-x} and \code{x} are assigned the middle color in the
color gradient. Note that values at exactly \code{x} or \code{-x}
are assigned the next respective color away from the middle
color. When \code{floor=0} or \code{floor=NULL} no floor is applied,
and colors are assigned using a continuous range of
numeric values from \code{-x} to \code{x} with length \code{n}.}

\item{baseline}{\code{numeric} value to define the baseline value, used
when zero is not the initial value. Note that \code{baseline} can be
either higher or lower than \code{x}, and colors from \code{colramp} will
be applied starting at \code{baseline} through \code{x}.}

\item{lens}{\code{numeric} value indicating a color lens applied
to the color gradient, passed to \code{jamba::getColorRamp()}.
Lens values \code{lens > 0} will condense the color gradient,
making smaller changes more visually distinct; \code{lens < 0}
expands the color gradient, making smaller changes less
visually distinct.}

\item{n}{\code{integer} number of colors used for the initial
color gradient. This value is forced to be an odd number,
so the "middle color" will always be represented as one
strict color. Note that when using a \code{floor}, the first
non-middle color is used for the \code{floor} assignment
which means a smaller \code{n} value will assign a more visibly
distinct color than using a larger \code{n}. See examples.}

\item{colramp}{\code{character} passed to \code{jamba::getColorRamp()}
which recognizes one of several forms of input:
\itemize{
\item \code{character} string matching the name of a color ramp
from \code{RColorBrewer} (see divergent palettes with
\code{RColorBrewer::display.brewer.all(type="div")}).
Note that adding \code{"_r"} will reverse the color gradient,
so the default \code{"BuRd_r"} will create a color gradient
with "blue-white-red" - with red for high values
consistent with "heat" in "heatmaps" - where heat is red.
\item \code{character} vector of R colors, which define a specific
color ramp. This vector will be expanded to \code{n} length.
}}

\item{open_floor}{\code{logical} indicating whether colors below
the assigned \code{floor} will still receive non-middle color.
Setting \code{open_floor=TRUE} is the best method to compare
the effect of assigning the strict middle-color to values
below the \code{floor}, versus using gradient colors below the
\code{floor}, while all remaining numeric-color assignments
are held constant.}

\item{debug}{\code{logical} indicating whether to produce a plot
that shows the resulting color gradient.}

\item{...}{additional arguments are ignored.}
}
\description{
Linear color interpolation function with adjustable range, baseline, and floor
}
\details{
This function is the linear equivalent of \code{col_div_xf()}, in that
it takes linear/sequential color gradient instead of a divergent
color gradient.
}
\examples{
col_fn1 <- col_linear_xf(x=3, baseline=0, floor=0)
col_fn2 <- col_linear_xf(x=3, baseline=0, floor=1)

col_fn2o <- col_linear_xf(x=3, baseline=0, floor=1, open_floor=TRUE)

test_seq <- seq(from=0, to=3, by=0.05);
names(test_seq) <- round(test_seq, digits=2);

opar <- par("mfrow"=c(1, 1));
bp0 <- barplot(abs(test_seq),
   las=2, yaxt="n",
   main="floor=0",
   col=col_fn1(test_seq),
   border="#22222222")
abline(v=bp0[abs(test_seq) == 1,], lty="dashed")
bp1 <- barplot(abs(test_seq),
   las=2, yaxt="n",
   main="floor=1",
   col=col_fn2(test_seq),
   border="#22222222")
abline(v=bp1[abs(test_seq) == 1,], lty="dashed")
bp2 <- barplot(abs(test_seq),
   las=2, yaxt="n",
   main="floor=1\nopen_floor=TRUE",
   col=col_fn2o(test_seq),
   border="#22222222")
abline(v=bp2[abs(test_seq) == 1,], lty="dashed")
par(opar)

col_fn3 <- col_linear_xf(x=3, baseline=6, floor=5)
test_seq <- seq(from=0, to=7, by=0.1);
names(test_seq) <- round(test_seq, digits=2);
bp3 <- barplot(abs(test_seq),
   las=2, yaxt="n",
   main="baseline=6, x=3, floor=5",
   col=col_fn3(test_seq),
   border="#22222222")
abline(v=bp3[abs(test_seq) == 5,], lty="dashed")
abline(v=bp3[abs(test_seq) == 3,], lty="dashed")

}
\seealso{
Other colorjam gradients: 
\code{\link{col_div_xf}()},
\code{\link{make_jam_divergent}()},
\code{\link{twostep_gradient}()}

Other colorjam assignment: 
\code{\link{col_div_xf}()},
\code{\link{group2colors}()},
\code{\link{matrix2heatColors}()},
\code{\link{vals2colorLevels}()}
}
\concept{colorjam assignment}
\concept{colorjam gradients}
