% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-add-colors.R
\name{color_distance}
\alias{color_distance}
\title{Calculate color distance between two or more colors}
\usage{
color_distance(
  x,
  y = NULL,
  method = c("cie2000", "cie94", "cie1976", "cmc", "euclidean"),
  ...
)
}
\arguments{
\item{x}{\code{character} color, required input colors.
\itemize{
\item When \code{y} is supplied, values in \code{y} are recycled to \code{length(x)},
and each entry in \code{x} is directly compared to \code{y}.
\item When \code{y} is not supplied, \code{x} is compared to itself,
returning a \code{matrix}.
}}

\item{y}{\code{character} color, default NULL}

\item{...}{additional arguments are passed to \code{farver::compare_colour()}}
}
\value{
\code{numeric} color distance with \code{length(x)} entries when both
\code{x} and \code{y} are supplied, or a \code{numeric} matrix with color distances
between all entries in \code{x}.
}
\description{
Calculate color distance between two or more colors
}
\details{
Color distance is calculated using \code{farver::compare_colour()},
with some defaults intended in future to assist with color blindness
calculations.
}
\examples{
color_distance("red", "firebrick")
color_distance("red", "red2")

color_distance(grDevices::palette.colors(15))

pc <- grDevices::palette.colors(15);
pc <- rainbowJam(5)
cd <- color_distance(pc);
cd[lower.tri(cd)] <- NA;
cse <- SummarizedExperiment::SummarizedExperiment(
   assays=list(distance=cd),
   rowData=data.frame(color=colnames(cd)),
   colData=data.frame(color=colnames(cd)))
jamses::heatmap_se(cse, use_raster=FALSE,
   top_colnames="color", rowData_colnames="color",
   sample_color_list=list(color=pc),
   cluster_columns=FALSE, cluster_rows=FALSE, color_max=360,
   centerby_colnames=FALSE,
   legend_at=seq(0, 300, by=100), legend_labels=seq(0, 300, by=100))

}
