% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-shiny.R
\name{colorjamShinyUI}
\alias{colorjamShinyUI}
\title{Colorjam R-shiny app UI}
\usage{
colorjamShinyUI(...)
}
\arguments{
\item{...}{additional arguments are ignored.}
}
\value{
A user interface definition suitable to pass
to \code{shiny::shinyUI()}.
}
\description{
Colorjam R-shiny app UI
}
\details{
This function contains the UI for the Colorjam R-shiny app.

The Colorjam R-shiny app is started by \code{launchColorjamShiny()}, which
calls \code{shiny::shinyApp()}, using arguments \code{server}, \code{ui},
\code{onStart}, and \code{options}. This function fulfills the
argument \code{ui}.
}
\concept{colorjam R-shiny functions

#@import shiny
#@import shinydashboard
#@import shinydashboardPlus
#@import shinyWidgets}
