% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-subset.R
\name{colors_to_df}
\alias{colors_to_df}
\title{Convert colors to data.frame of color attributes}
\usage{
colors_to_df(x, ..., alpha = TRUE, byCols = NULL)
}
\arguments{
\item{x}{\code{character} vector of R colors}

\item{...}{any valid criteria to subset the color \code{data.frame}.}

\item{alpha}{\code{logical} indicating whether to retain \code{alpha}
transparency in intermediate operations. When \code{alpha=FALSE} then
all color transparency is ignored, and colors will be returned
with no transparency.}

\item{byCols}{\code{character} vector of colnames to sort after the subset
operation.}
}
\value{
\code{data.frame} with color attributes as columns.
When \code{byCols} is defined, the \code{data.frame} is sorted using
\code{jamba::mixedSortDF()}.
When \code{...} is supplied, \code{subset()} is applied to subset colors.
}
\description{
Convert colors to data.frame of color attributes
}
\details{
This function takes a vector of colors and returns a \code{data.frame}
with relevant color attributes:
\itemize{
\item \code{"num"} contains the integer index of the input vector \code{x}.
\item \code{"hex"} contains \code{character} values with hexadecimal colors
including alpha. If the input \code{x} is \code{"red"} then the hex value
will be converted to \code{"#FF0000FF"}.
\item \code{"name"} if \code{names(x)} is not empty
}
}
\section{Added by \code{jamba::col2hcl()}}{
\itemize{
\item \code{"H"} contains color hue as values from \code{0} to \code{360}.
\item \code{"C"} contains color chroma (aka saturation) ranging from \code{0} up to \code{200},
where typical "full saturation" is represented as values above 100.
\item \code{"L"} contains color luminance (brightness/lightness) ranging from
\code{0} to \code{100}.
\item \code{"alpha"} the alpha transparency, ranging from \code{0} (fully transparent)
to \code{1} (fully opaque, not transparent).
}
}

\section{Added by \code{grDevices::col2rgb()}}{
\itemize{
\item \code{"red"} contains the red color channel, values range from \code{0} to \code{255}.
\item \code{"green"} contains the green color channel, values range from \code{0} to \code{255}.
\item \code{"blue"} contains the blue color channel, values range from \code{0} to \code{255}.
}
}

\section{Added by \code{jamba::col2hsv()}}{
\itemize{
\item \code{"h"} contains color hue as values from \code{0} to \code{1}. Note these values
may not map directly to color hue obtained from \code{jamba::col2hcl()}.
\item \code{"s"} contains color saturation with values from \code{0} to \code{1}.
\item \code{"v"} contains color vibrance (brightness/lightness) with values
from \code{0} to \code{1}.
}
}

\section{Added by \code{jamba::col2hsl()}}{
\itemize{
\item \code{"hsl_h"} contains color hue as values from \code{0} to \code{1}. Note these values
may not map directly to color hue obtained from \code{jamba::col2hcl()}.
\item \code{"hsl_s"} contains color saturation with values from \code{0} to \code{100}.
\item \code{"hsl_l"} contains color luminance (brightness/lightness) with values
from \code{0} to \code{100}.
}
}

\seealso{
Other colorjam sort: 
\code{\link{sort_colors}()},
\code{\link{subset_colors}()}

Other colorjam core: 
\code{\link{blend_colors}()},
\code{\link{closestRcolor}()},
\code{\link{group2colors}()},
\code{\link{rainbowJam}()},
\code{\link{sort_colors}()},
\code{\link{subset_colors}()}
}
\concept{colorjam core}
\concept{colorjam sort}
