% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-blendcolors.R
\name{combine_alphas}
\alias{combine_alphas}
\title{Combine alpha transparency values with additive logic}
\usage{
combine_alphas(x, max_alpha = 1, ...)
}
\arguments{
\item{x}{\code{numeric} alpha value, typically limited between 0 and 1.
The max value can be set using \code{max_alpha}.}

\item{max_alpha}{\code{numeric} default 1, the maximum permitted alpha value.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{numeric} value after combining alpha values.
}
\description{
Combine alpha transparency values with additive logic
}
\details{
Alpha transparency is defined as 0 for fully transparent, and 1
(or \code{max_alpha}) for fully opaque (not transparent).
The purpose is to permit combining multiple colors, where the
alpha transparency builds over time. Each color should contribute
some proportional fraction to the overall opacity of the final
color.

The basic formula:

\code{new_alpha <- alpha1 + (1 - alpha1) * alpha2}

Or when \code{max_alpha} is defined:

\code{new_alpha <- alpha1 + (max_alpha - alpha1) * alpha2}
\itemize{
\item Any \code{NA} values are considered equivalent to \code{0} and are
therefore not applied.
\item All input alpha values are restricted to values between
\code{0} and \code{max_alpha}.
\item For more than two values, each value is applied in series,
which works out to the same result if applied in any order.
}
}
\examples{
# it progressively fills 50\% of remaining transparency
combine_alphas(c(0.5, 0.5))
combine_alphas(c(0.5, 0.5, 0.5))
combine_alphas(c(0.5, 0.5, 0.5, 0.5))

base_alpha <- 0.5;
new_alphas <- sapply(1:5, function(i){
   combine_alphas(rep(base_alpha, i))
})
names(new_alphas) <- seq_along(new_alphas);
bp <- barplot(new_alphas, ylim=c(0, 1.1), col="navy",
   xlab=paste0("Number of ",
   base_alpha,
   " alpha values combined"))
abline(h=1, lty=2);
jamba::shadowText(x=bp[, 1], y=new_alphas,
   col="white", cex=1.5,
   pos=1,
   xpd=TRUE,
   label=round(new_alphas, digits=3))
}
\seealso{
Other colorjam display: 
\code{\link{color_pie}()},
\code{\link{showDichromat}()}
}
\concept{colorjam display}
