% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-angles.R
\name{display_degrees}
\alias{display_degrees}
\title{Display degree angles around a unit circle}
\usage{
display_degrees(
  x,
  add = FALSE,
  col = "darkorange",
  lwd = 2,
  top_degree = 0,
  clockwise = TRUE,
  r1 = 0,
  r2 = 1,
  ...
)
}
\description{
Display degree angles around a unit circle
}
\examples{
display_degrees(c(0, 45), col=c("red", "blue"))

}
\seealso{
Other colorjam hue warp: 
\code{\link{approx_degrees}()},
\code{\link{h2hwOptions}()},
\code{\link{h2hw}()},
\code{\link{hw2h}()},
\code{\link{mean_angle}()}
}
\concept{colorjam hue warp}
