% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-color.R
\name{group2colors}
\alias{group2colors}
\title{Assign colors to vector of group labels}
\usage{
group2colors(x, alpha = 1, colorFunc = rainbowJam,
  useGradient = FALSE, verbose = FALSE, ...)
}
\description{
Assign colors to vector of group labels
}
\details{
This function takes a character or factor vector as input, then
assigns categorical colors to each label using \code{colorFunc}, by
default \code{rainbowJam()}.
}
