% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-remap-colors.R
\name{hsl_to_hcl_hue}
\alias{hsl_to_hcl_hue}
\title{Convert HSL hue to HCL hue}
\usage{
hsl_to_hcl_hue(x, ...)
}
\arguments{
\item{x}{\code{numeric} hue with values in range \code{c(0, 360)}
intended for use with \code{jamba::hsl2col()}}

\item{...}{additional arguments are ignored.}
}
\value{
\code{numeric} hue with values in range \code{c(0, 360)}
intended for use with \code{jamba::hcl2col()}
}
\description{
Convert HSL hue to HCL hue
}
\seealso{
Other colorjam hue warp: 
\code{\link{add_colorjam_preset}()},
\code{\link{add_colorjam_step}()},
\code{\link{adjust_hue_warp}()},
\code{\link{approx_degrees}()},
\code{\link{colorjam_presets}()},
\code{\link{colorjam_steps}()},
\code{\link{display_degrees}()},
\code{\link{h2hwOptions}()},
\code{\link{h2hw}()},
\code{\link{hcl_to_hsl_hue}()},
\code{\link{hw2h}()},
\code{\link{mean_angle}()},
\code{\link{plot_colorjam_preset}()},
\code{\link{remap_colorjam_preset}()},
\code{\link{validate_colorjam_preset}()}
}
\concept{colorjam hue warp}
