% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-objects.R
\docType{data}
\name{jam_divergent}
\alias{jam_divergent}
\title{Jam Divergent Color Gradients}
\format{An object of class \code{list} of length 14.}
\usage{
jam_divergent
}
\description{
Jam Divergent Color Gradients that are color-blind-friendly.
}
\details{
This data contains a \code{list} of divergent color gradients generated
similar to \code{jam_linear} with multi-hue technique based upon
\code{RColorBrewer}. Each gradient is designed so that
the two colors are visibly distinct even under three
different color blindness simulations from \code{dichromat::dichromat()}.

Each gradient is named by the closest corresponding R colors,
where the same names are used in \code{jam_linear} so that colors
can be matched where appropriate:
\itemize{
\item skyblue_firebrick
\item dodgerblue_orangered
\item royalblue_tomato
\item slateblue_sienna
\item orchid_orange
\item magenta_goldenrod
\item maroon_gold
\item firebrick_skyblue
\item orangered_dodgerblue
\item tomato_royalblue
\item sienna_slateblue
\item orange_orchid
\item goldenrod_magenta
\item gold_maroon
}

These linear gradients are designed with a white background
color.
These divergent gradients are designed with
a black background, intended to help indicate that these
colors are divergent. Note that linear gradients \code{jam_linear}
are designed with a white background color.
}
\examples{
data(jam_divergent)
jamba::showColors(jam_divergent)

# show the same with dichromat adjustment
if (require(dichromat)) {
   jamba::showColors(lapply(jam_divergent, dichromat::dichromat))
}

}
\seealso{
Other colorjam data: 
\code{\link{jam_linear}}
}
\concept{colorjam data}
\keyword{datasets}
