% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-objects.R
\docType{data}
\name{jam_linear}
\alias{jam_linear}
\title{Jam Linear Color Gradients}
\format{
An object of class \code{list} of length 14.
}
\usage{
jam_linear
}
\description{
Jam Linear Color Gradients that are color-blind-friendly.
}
\details{
This data contains a \code{list} of linear/sequential color
gradients generated with
a multi-hue technique similar to that used by \code{RColorBrewer}
to expand the visual clarity of a linear single-hue color
gradient. Colors were chosed to avoid common color-blindness
problems, and to maximize visual differences between
color gradients.

The colors in \code{jam_linear} are designed so that colors
in \code{jam_divergent} can be chosen for consistency.
For example linear color gradient \code{"firebrick"} can
be used to show coverage values, and \code{"skyblue_firebrick"}
can be used to show coverage difference from control,
and in both cases the maximum color is \code{"firebrick"}.

These linear gradients are designed with a white background
color. Note the \code{jam_divergent} colors are designed with
a black background, intended to help indicate that these
colors are divergent.

In general, there are seven warm color tones, and seven cool color
tones. Visual distinction is only expected between warm and
cool color tones, and is not distinct within the same color
tone.

Each color is named by the closest corresponding R color:
\itemize{
\item firebrick
\item orangered
\item tomato
\item sienna
\item goldenrod
\item gold
\item skyblue
\item dodgerblue
\item royalblue
\item slateblue
\item orchid
\item magenta
\item maroon
}
}
\examples{
data(jam_linear)
jamba::showColors(jam_linear)

# show the same with dichromat adjustment
if (require(dichromat)) {
   jamba::showColors(lapply(jam_linear, dichromat::dichromat))
}

}
\seealso{
Other colorjam data: 
\code{\link{jam_divergent}},
\code{\link{named_colors}}
}
\concept{colorjam data}
\keyword{datasets}
