% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-ggplot2-themejam.R
\name{jam_pal}
\alias{jam_pal}
\title{Jam color palette for ggplot2}
\usage{
jam_pal(
  type = "seq",
  palette = 1,
  direction = 1,
  invert = FALSE,
  darkFactor = 1,
  sFactor = 1,
  darken = FALSE,
  alpha = 1,
  useGrey = 20,
  preset = getOption("colorjam.preset", "dichromat2"),
  step = getOption("colorjam.step", "default"),
  ...
)
}
\arguments{
\item{type}{\code{character} string indicating the colors are sequential
\code{"seq"}.}

\item{palette}{\code{integer} value indicating the categorical palette
to use, intended to provide variety in the color assignment.
(Not yet implemented.)}

\item{direction}{\code{integer} indicating whether to reverse the color
assignment, either \code{1} for the default forward assignment, or
\code{-1} for reverse assignment. Any negative value will reverse
the colors.}

\item{invert}{\code{logical} indicating whether to return corresponding
contrasting colors, for example for text labels, typically either
\code{"white"} or \code{"black"} as defined by \code{jamba::setTextContrastColor()}.}

\item{darkFactor, sFactor}{\code{numeric} passed to \code{jamba::makeColorDarker()}
for optional adjustment of the color, by darkness and saturation,
respectively. Intended when using \code{scale_color_jam()} and
\code{scale_fill_jam()} where you want the color value to be lighter
or darker than the fill color, a useful effect for outlines.}

\item{darken}{\code{logical} indicating whether to apply default values
for \code{darkFactor} and \code{sFactor} to darken the resulting colors.}

\item{alpha}{\code{numeric} value indicating the alpha transparency, on a
scale of 0 (transparent) to 1 (non-transparent).}

\item{useGrey}{\code{integer} value between 0 and 100 indicating the grey
value, as sent to \code{jamba::setTextContrastColor()}, used only when
\code{invert=TRUE}.}

\item{...}{additional arguments are passed to \code{rainbowJam()}.}
}
\description{
Jam color palette for ggplot2
}
\seealso{
Other colorjam ggplot2: 
\code{\link{scale_color_jam}()},
\code{\link{scale_fill_jam}()},
\code{\link{theme_jam}()}
}
\concept{colorjam ggplot2}
