% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-shiny.R
\name{launchColorjamShiny}
\alias{launchColorjamShiny}
\title{Colorjam R-shiny App}
\usage{
launchColorjamShiny(
  ...,
  envir = globalenv(),
  options = list(width = 1200, host = "0.0.0.0", port = 8080),
  verbose = TRUE
)
}
\arguments{
\item{...}{additional arguments are passed to \code{shiny::shinyApp()}
when the argument name is recognized by \code{shiny::shinyApp()}.}

\item{envir}{\code{environment}}

\item{options}{\code{list} passed to \code{shiny::shinyApp()}. The most common
options are:
\itemize{
\item \code{host} with either a specific IP address which then requires
all incoming web requests to match that IP address,
or \code{host="0.0.0.0"} to accept web requests for any host.
\item \code{port} with specific \code{numeric} web port, or \code{port=NULL}
to assign a random port number when the R-shiny app is started.
\item \code{width} with a user-configurable width for the R-shiny app,
useful when graphical components should expect at least a minimum
width.
}}

\item{verbose}{\code{logical} indicating whether to print verbose output.}
}
\value{
An object that represents the R-shiny app. Printing the
object or passing it to \code{shiny::runApp()} will run the app.
}
\description{
Colorjam R-shiny App for color selection and custom adjustments
}
\examples{
if (interactive()) {
   launchColorjamShiny();
}
}
\seealso{
Other colorjam R-shiny functions: 
\code{\link{colorjamShinyServer}()}
}
\concept{colorjam R-shiny functions}
