% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-blendcolors.R
\name{mean_angle}
\alias{mean_angle}
\title{Calculate the mean angle}
\usage{
mean_angle(x, w = NULL, do_plot = FALSE, lens = -5, ...)
}
\arguments{
\item{x}{\code{numeric} vector of angles in degrees}

\item{w}{\code{numeric} vector representing weights}

\item{do_plot}{\code{logical} indicating whether to create a visual summary plot}

\item{lens}{\code{numeric} value passed to \code{jamba::warpAroundZero()} to adjust
the radius}

\item{...}{additional arguments are ignored}
}
\value{
\code{numeric} vector that contains
\itemize{
\item \code{degree} the mean angle in degrees
\item \code{radius} the actual radius based upon mean unit vectors
\item \code{radius2} the adjusted radius using \code{jamba::warpAroundZero()}
}
}
\description{
Calculate the mean angle
}
\details{
This function takes a vector of angles in degrees (0 to 360 degrees)
and returns the mean angle based upon the average of unit vectors.

The function also optionally accomodates weighted mean values,
if a vector of weights is supplied as \code{w}.

Part of the intent of this function is to be used for color blending
methods, for example taking the average color hue from a vector of
colors. For this purpose, some colors may have varying color saturation
and transparency, which are mapped here as weight \code{w}. Colors which are
fully transparent should therefore have weight \code{w=0} so they do not
contribute to the resulting average color hue. Also during color blending
operations, the resulting color saturation is adjusted using the \code{lens}
argument, the default \code{lens=-5} has a tendency to increase intermediate
color saturation.
}
\examples{
mean_angle(c(120, 45), do_plot=TRUE);

}
\seealso{
Other colorjam hue warp: 
\code{\link{add_colorjam_preset}()},
\code{\link{add_colorjam_step}()},
\code{\link{adjust_hue_warp}()},
\code{\link{approx_degrees}()},
\code{\link{colorjam_presets}()},
\code{\link{colorjam_steps}()},
\code{\link{display_degrees}()},
\code{\link{h2hw}()},
\code{\link{h2hwOptions}()},
\code{\link{hcl_to_hsl_hue}()},
\code{\link{hsl_to_hcl_hue}()},
\code{\link{hw2h}()},
\code{\link{plot_colorjam_preset}()},
\code{\link{remap_colorjam_preset}()},
\code{\link{validate_colorjam_preset}()}
}
\concept{colorjam hue warp}
