% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-objects.R
\docType{data}
\name{named_colors}
\alias{named_colors}
\title{named_colors from meodai/color-names and R colors()}
\format{
a \code{character} vector with length 4883, containing
hexadecimal colors named with human-assigned color names.
\subsection{Processing}{

Named colors were obtained from the amazing Github repository
\href{https://github.com/meodai/color-names}{meodai/color-names}.
This resource provided 30200 total named colors.
A subset of 4447 color names were denoted "good names" and were
included here.

R colors from \code{grDevices::colors()} whose hex values were not already
included in the Meodai resource were added, after removing identical
hex colors such as \code{"darkred"} and \code{"red4"}.
These R colors were named with prefix \code{"R"} so labels such as
\code{"darkred"} became \code{"Rdarkred"}. The \code{"R"} prefix was used to avoid
name clashes, and to permit the Meodai color names to take precedence
whenever the same name was used in both resources.
The Meodai resource seemed visually preferable to similarly-named
R \code{grDevices::colors()} such as \code{"brown"}, and \code{"chartreuse"},
however there could be benefit in using R color names since they
appear to match the W3C color name standard used by HTML and CSS.
}
}
\source{
\url{https://github.com/meodai/color-names}
}
\usage{
named_colors
}
\description{
named_colors including 4447 entries from meodai/color-names,
and 436 entries from R \code{colors()}.
}
\examples{
jamba::showColors(colorjam::named_colors)

# subset for color chroma at least 30
colorjam::color_pie(unname(
   colorjam::subset_colors(colorjam::named_colors, C >= 30)))

}
\seealso{
Other colorjam data: 
\code{\link{jam_divergent}},
\code{\link{jam_linear}}
}
\concept{colorjam data}
\keyword{datasets}
