% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-presets.R
\name{plot_colorjam_preset}
\alias{plot_colorjam_preset}
\title{Plot colorjam preset}
\usage{
plot_colorjam_preset(
  preset = NULL,
  h1 = NULL,
  h2 = NULL,
  direction = 1,
  xlim = c(0, 360),
  ylim = c(0, 360),
  digits = 4,
  cex = 1,
  col = "#FFFFFF",
  pt_pch = 20,
  pt_cex = 1,
  pt_col = "blue",
  pt_step = 5,
  colorize_borders = TRUE,
  axis_fontfamily = "Arial",
  axislabel_fontsize = 24,
  axis_fontsize = 18,
  style = c("base", "plotly"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{colorize_borders}{\code{logical} indicating whether to display the color
spectrum along the x-axis and y-axis borders.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Plot colorjam preset
}
\details{
Plot the \code{h1},\code{h2} color hue association, by default shown
relative to the \code{h2} (virtual hue) value on the x-axis,
and \code{h1} (actual hue) on the y-axis.

Any duplicated \code{h2} values are indicated with a vertical line.
}
\seealso{
Other colorjam hue warp: 
\code{\link{add_colorjam_preset}()},
\code{\link{add_colorjam_step}()},
\code{\link{adjust_hue_warp}()},
\code{\link{approx_degrees}()},
\code{\link{colorjam_presets}()},
\code{\link{colorjam_steps}()},
\code{\link{display_degrees}()},
\code{\link{h2hw}()},
\code{\link{h2hwOptions}()},
\code{\link{hcl_to_hsl_hue}()},
\code{\link{hsl_to_hcl_hue}()},
\code{\link{hw2h}()},
\code{\link{mean_angle}()},
\code{\link{remap_colorjam_preset}()},
\code{\link{validate_colorjam_preset}()}
}
\concept{colorjam hue warp}
