% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-steps.R
\name{plot_colorjam_steps}
\alias{plot_colorjam_steps}
\title{Plot colorjam steps}
\usage{
plot_colorjam_steps(
  step,
  n = NULL,
  hue = NULL,
  lines = TRUE,
  return_type = c("ptcol", "step_data"),
  ...
)
}
\arguments{
\item{step}{\code{character} name of colorjam step from \code{colorjam_steps()}}

\item{n}{\code{integer} number of steps to show, default NULL uses all steps.}

\item{hue}{\code{numeric} hue or \code{character} color to convert to color hue,
used to colorize points using the actual C,L step values.
\itemize{
\item When no \code{hue} is supplied, the points are colored using a color
gradient \code{"Reds"} in the order they appear.
}}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} with step values, invisibly.
}
\description{
Plot colorjam steps in simple scatterplot format
}
\details{
Simple function to plot the series of Chroma/Luminance steps
for a given named step series. See steps with \code{colorjam_steps()}.
}
\examples{
colorjam_steps()
step_name <- colorjam_steps()[1]
plot_colorjam_steps(step_name)

step_name <- tail(colorjam_steps(), 1)
plot_colorjam_steps(step_name)

plot_colorjam_steps(step_name, hue="orange")
plot_colorjam_steps(step_name, hue="gold")

}
