% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-rainbowjammulti.R
\name{rainbowJamMulti}
\alias{rainbowJamMulti}
\title{Rainbow categorical colors across multiple groups}
\usage{
rainbowJamMulti(ns, do_plot = FALSE, ...)
}
\arguments{
\item{ns}{\code{integer} vector of \code{n} values passed to
\code{rainbowJam()}.}

\item{do_plot}{\code{logical} indicating whether to plot the result
using \code{jamba::showColors()}.}

\item{...}{additional arguments are passed to \code{rainbowJam()}.}
}
\value{
\code{list} of categorical colors, length equal to the
input \code{length(ns)}, with color vectors with lengths \code{ns}.
}
\description{
Rainbow categorical colors across multiple groups
}
\details{
This function applies \code{colorjam::rainbowJam()} to multiple
vectors, such that the categorical colors are not duplicated,
and are assigned as well-spaced across the rainbow for
each group as possible.
}
\examples{
opar <- par("mfrow"=c(3, 1));
on.exit(par(opar));
ns <- c(A=8, B=3, C=4);
colorlist1 <- rainbowJamMulti(ns,
   do_plot=TRUE,
   main="rainbowJamMulti()");

# basic assignment in order
colorset2 <- colorjam::rainbowJam(sum(ns));
colorlist2 <- split(colorset2, rep(seq_along(ns), ns))
jamba::showColors(colorlist2, main="sequential assignment")

# re-assign the same colors
colorlist3 <- lapply(ns, colorjam::rainbowJam);
jamba::showColors(colorlist3, main="independent color assignment");
par(opar);

}
\seealso{
Other colorjam assignment: 
\code{\link{col_div_xf}()},
\code{\link{col_linear_xf}()},
\code{\link{group2colors}()},
\code{\link{matrix2heatColors}()},
\code{\link{vals2colorLevels}()}
}
\concept{colorjam assignment}
