% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-remap-colors.R
\name{remap_colorjam_preset}
\alias{remap_colorjam_preset}
\title{Remap colors to a new colorjam preset}
\usage{
remap_colorjam_preset(
  x,
  preset = NULL,
  preset_from = NULL,
  do_plot = FALSE,
  plot_debug = FALSE,
  ...
)
}
\value{
\code{character} vector of hexadecimal colors
}
\description{
Remap colors to a new  colorjam preset, using existing preset
when available.
}
\details{
This function is experimental, and is intended to convert a set
of categorical colors to a colorjam preset. When the existing
colors have attribute \code{"preset"} it is used to inform the starting
preset. The output of \code{rainbowJam()} includes the preset as
\code{attr(x, "preset")}.

The HCL and HSL color conversions are blended together to improve
the imperfect result from either method alone. The end result
is imperfect, but better than the alternatives.
}
\examples{
x <- rainbowJam(12)
x_new <- remap_colorjam_preset(x, preset="ryb2", do_plot=TRUE)

x <- rainbowJam(12, preset="ryb2")
x_new <- remap_colorjam_preset(x, preset="dichromat2", do_plot=TRUE)

x <- rainbowJam(12, preset="ryb")
x_new <- remap_colorjam_preset(x, preset="dichromat2", do_plot=TRUE)

x <- rainbowJam(12, preset="dichromat2")
x_new <- remap_colorjam_preset(x, preset="ryb", do_plot=TRUE)

x <- rainbowJam(12, preset="rgb")
x_new <- remap_colorjam_preset(x, preset="ryb", do_plot=TRUE)

}
\seealso{
Other colorjam hue warp: 
\code{\link{add_colorjam_preset}()},
\code{\link{add_colorjam_step}()},
\code{\link{adjust_hue_warp}()},
\code{\link{approx_degrees}()},
\code{\link{colorjam_presets}()},
\code{\link{colorjam_steps}()},
\code{\link{display_degrees}()},
\code{\link{h2hw}()},
\code{\link{h2hwOptions}()},
\code{\link{hcl_to_hsl_hue}()},
\code{\link{hsl_to_hcl_hue}()},
\code{\link{hw2h}()},
\code{\link{mean_angle}()},
\code{\link{plot_colorjam_preset}()},
\code{\link{plot_colorjam_steps}()},
\code{\link{validate_colorjam_preset}()}
}
\concept{colorjam hue warp}
