% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-color.R
\name{scale_color_jam}
\alias{scale_color_jam}
\title{Apply rainbowJam categorical colors to a ggplot2 object}
\usage{
scale_color_jam(..., type = "seq", palette = 1, direction = 1,
  invert = FALSE, useGrey = 20)
}
\arguments{
\item{...}{additional arguments are passed to \code{ggplot2::discrete_scale()}.}

\item{type}{character string indicating the colors are sequential
\code{"seq"}, and is passed to \code{colorjam::jam_pal()}.}

\item{palette}{integer value indicating the categorical palette
to use, intended to provide variety in the color assignment.
(Not yet implemented.)}

\item{direction}{integer indicating whether to reverse the color
assignment, either \code{1} for the default forward assignment, or
\code{-1} for reverse assignment. Any negative value will reverse
the colors.}

\item{invert}{logical indicating whether to return corresponding
contrasting colors, for example for text labels, typically either
\code{"white"} or \code{"black"} as defined by \code{jamba::setTextContrastColor()}.}

\item{useGrey}{integer value between 0 and 100 indicating the grey
value, as sent to \code{jamba::setTextContrastColor()}, used only when
\code{invert=TRUE}.}
}
\description{
Apply rainbowJam categorical colors to a ggplot2 object
}
\details{
This function provides a function in the format \code{scale_color_*}
to be applied to ggplot2 objects. It can provide a more visibly
distinct set of categorical colors than \code{ggplot2::scale_color_hue()}.
}
\examples{
if (suppressPackageStartupMessages(require(ggplot2))) {
   dsamp <- diamonds[sample(nrow(diamonds), 1000),];
   (d <- ggplot(dsamp, aes(carat, price)) + geom_point(aes(colour=clarity)));

   d + scale_color_hue() + ggtitle("scale_color_hue()");
   d + scale_color_jam() + ggtitle("scale_color_jam()");
}
}
