% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-scale-color-div-xf.R
\name{scale_color_linear_xf}
\alias{scale_color_linear_xf}
\alias{scale_fill_linear_xf}
\title{Continuous linear color scale for ggplot2, with max and optional floor}
\usage{
scale_color_linear_xf(
  x = 1,
  floor = 0,
  baseline = 0,
  lens = 0,
  n = 6,
  colramp = "Purples",
  open_floor = FALSE,
  floor_color = NULL,
  guide = NULL,
  breaks = NULL,
  hjust = 1,
  aesthetics = "color",
  ...
)

scale_fill_linear_xf(
  x = 1,
  floor = 0,
  baseline = 0,
  lens = 0,
  n = 6,
  colramp = "Purples",
  open_floor = FALSE,
  floor_color = NULL,
  guide = NULL,
  breaks = NULL,
  hjust = 1,
  aesthetics = "fill",
  ...
)
}
\arguments{
\item{x}{\code{numeric} value used as a threshold, where numeric
values at or above this value \code{x} are assigned the last
color in the color gradient. Negative values at or below
this negative value \code{-x} are assigned the first color
in the color gradient.}

\item{floor}{\code{numeric} optional value where numeric values
between \code{-x} and \code{x} are assigned the middle color in the
color gradient. Note that values at exactly \code{x} or \code{-x}
are assigned the next respective color away from the middle
color. When \code{floor=0} or \code{floor=NULL} no floor is applied,
and colors are assigned using a continuous range of
numeric values from \code{-x} to \code{x} with length \code{n}.}

\item{baseline}{\code{numeric} value to define the baseline value, used
when zero is not the initial value. Note that \code{baseline} can be
either higher or lower than \code{x}, and colors from \code{colramp} will
be applied starting at \code{baseline} through \code{x}.}

\item{lens}{\code{numeric} value indicating a color lens applied
to the color gradient, passed to \code{jamba::getColorRamp()}.
Lens values \code{lens > 0} will condense the color gradient,
making smaller changes more visually distinct; \code{lens < 0}
expands the color gradient, making smaller changes less
visually distinct.}

\item{n}{\code{integer} number of colors used for the initial
color gradient. This value is forced to be an odd number,
so the "middle color" will always be represented as one
strict color. Note that when using a \code{floor}, the first
non-middle color is used for the \code{floor} assignment
which means a smaller \code{n} value will assign a more visibly
distinct color than using a larger \code{n}. See examples.}

\item{colramp}{\code{character} passed to \code{jamba::getColorRamp()}
which recognizes one of several forms of input:
\itemize{
\item \code{character} string matching the name of a color ramp
from \code{RColorBrewer} (see divergent palettes with
\code{RColorBrewer::display.brewer.all(type="div")}).
Note that adding \code{"_r"} will reverse the color gradient,
so the default \code{"BuRd_r"} will create a color gradient
with "blue-white-red" - with red for high values
consistent with "heat" in "heatmaps" - where heat is red.
\item \code{character} vector of R colors, which define a specific
color ramp. This vector will be expanded to \code{n} length.
}}

\item{open_floor}{\code{logical} indicating whether colors below
the assigned \code{floor} will still receive non-middle color.
Setting \code{open_floor=TRUE} is the best method to compare
the effect of assigning the strict middle-color to values
below the \code{floor}, versus using gradient colors below the
\code{floor}, while all remaining numeric-color assignments
are held constant.}

\item{floor_color}{\code{character} default NULL, optional fixed color
to apply below the floor threshold. Note that this color
works best when \code{open_floor=FALSE}.}

\item{guide}{\code{GuideLegend} object used by ggplot2 to customize the
guide displayed in the legend. Default simply sets the guide legend
values to be reversed, so that higher numeric values appear at
the top of the legend.}

\item{breaks}{default NULL uses the breaks and labels defined by
\code{col_div_xf()}, otherwise suggested \code{ggplot2::waiver()} to use its
default.}

\item{hjust}{\code{numeric} used to define horizontal justification of numeric
labels. Default 1 uses right-alignment.}

\item{...}{additional arguments are passed to \code{ggplot2::continuous_scale()}.}
}
\value{
\code{ggplot2::ScaleContinuous} object which inherits \code{ggproto},
suitable to use in ggplot2 composition.
}
\description{
Continuous linear color scale for ggplot2, with max and optional floor
}
\details{
This function is intended to provide ggplot2 color scales
using the same logic applied in \code{col_linear_xf()}.
Specifically, it applies a maximum numeric value to the color range,
and an optional numeric floor.

Other enhancements:
\itemize{
\item The ggplot2 legend shows discrete steps with labels for each color.
This style differs from ggplot2 "colorsteps" which labels the break
between colors.
}
}
\examples{
\dontshow{if ((requireNamespace("ggplot2", quietly=TRUE))) withAutoprint(\{ # examplesIf}
# ggplot2 is required
ggplot2::ggplot(ggplot2::faithfuld, ggplot2::aes(waiting, eruptions, fill=density)) +
   ggplot2::geom_raster() +
   scale_fill_linear_xf(0.04) +
   ggplot2::scale_x_continuous(expand = c(0, 0)) +
   ggplot2::scale_y_continuous(expand = c(0, 0)) +
   theme_jam()
\dontshow{\}) # examplesIf}
}
\seealso{
Other colorjam ggplot2: 
\code{\link{scale_color_div_xf}()},
\code{\link{scale_color_jam}()},
\code{\link{scale_fill_jam}()},
\code{\link{theme_jam}()}
}
\concept{colorjam ggplot2}
