% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-color.R
\name{scale_fill_jam}
\alias{scale_fill_jam}
\title{Apply rainbowJam categorical color fill to a ggplot2 object}
\usage{
scale_fill_jam(
  ...,
  type = "seq",
  palette = 1,
  direction = 1,
  invert = FALSE,
  darkFactor = 1,
  sFactor = 1,
  alpha = 1,
  useGrey = 20,
  preset = "dichromat"
)
}
\arguments{
\item{...}{additional arguments are passed to \code{ggplot2::discrete_scale()}.}

\item{type}{character string indicating the colors are sequential
\code{"seq"}, and is passed to \code{colorjam::jam_pal()}.}

\item{palette}{integer value indicating the categorical palette
to use, intended to provide variety in the color assignment.
(Not yet implemented.)}

\item{direction}{integer indicating whether to reverse the color
assignment, either \code{1} for the default forward assignment, or
\code{-1} for reverse assignment. Any negative value will reverse
the colors.}

\item{invert}{logical indicating whether to return corresponding
contrasting colors, for example for text labels, typically either
\code{"white"} or \code{"black"} as defined by \code{jamba::setTextContrastColor()}.}

\item{darkFactor, sFactor}{numeric sent to \code{jamba::makeColorDarker()} for
optional adjustment of the color, by darkness and saturation,
respectively. Intended when using \code{scale_color_jam()} and
\code{scale_fill_jam()} where you want the color value to be lighter
or darker than the fill color, a useful effect for outlines.}

\item{alpha}{numeric value indicating the alpha transparency, on a
scale of 0 (transparent) to 1 (non-transparent).}

\item{useGrey}{integer value between 0 and 100 indicating the grey
value, as sent to \code{jamba::setTextContrastColor()}, used only when
\code{invert=TRUE}.}
}
\description{
Apply rainbowJam categorical color fill to a ggplot2 object
}
\details{
This function provides a function in the format \verb{scale_fill_*}
to be applied to ggplot2 objects. It can provide a more visibly
distinct set of categorical colors than \code{ggplot2::scale_fill_hue()}.
}
\examples{
if (suppressPackageStartupMessages(require(ggplot2))) {
   dsamp <- ggplot2::diamonds[sample(nrow(ggplot2::diamonds), 1000),];
   d <- ggplot2::ggplot(dsamp, ggplot2::aes(carat, price)) +
    ggplot2::geom_point(ggplot2::aes(colour=cut, bg=cut), pch=21, size=4);

   print(d +
      scale_color_jam(darkFactor=1.5) +
      scale_fill_jam() +
      ggplot2::ggtitle("scale_color_jam(darkFactor=1.5) + scale_fill_jam()"));
}

}
\seealso{
Other colorjam ggplot2: 
\code{\link{jam_pal}()},
\code{\link{scale_color_jam}()},
\code{\link{theme_jam}()}
}
\concept{colorjam ggplot2}
