% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-color-distance.R
\name{show_color_distance}
\alias{show_color_distance}
\title{Show color distance as a heatmap}
\usage{
show_color_distance(
  cd = NULL,
  pc = NULL,
  show_labels = TRUE,
  cluster_data = FALSE,
  row_split = 0,
  ...
)
}
\arguments{
\item{cd}{\code{numeric} matrix, default NULL, with pre-calculated values.}

\item{pc}{\code{character} vector of colors used in \code{cd}, useful when
there are \code{names(pc)} to use as heatmap column or row labels.
When only \code{pc} is provided and not \code{cd}, the \code{cd} is calculated
by calling \code{color_distance(pc)}.}

\item{show_labels}{\code{logical} default TRUE, used when \code{pc} is provided.}

\item{cluster_data}{\code{logical} default FALSE, whether to enable row
and column hierarchical clustering in the heatmap.}

\item{row_split}{\code{numeric} default 0, used when \code{cluster_data=TRUE}
to subdivide the dendrogram into this many separate subclusters.}

\item{...}{additional arguments are passed to \code{color_distance()},
then to \code{jamses::heatmap_se()}. Many arguments in \code{jamses::heatmap_se()}
are also passed through to \code{ComplexHeatmap::Heatmap()}.}
}
\value{
\code{ComplexHeatmap::Heatmap} object, when printed it will
draw a heatmap.
}
\description{
Show color distance as a heatmap
}
\details{
This function uses \code{ComplexHeatmap::Heatmap()}, which is not
required for colorjam as a whole. It creates a matrix visual
summary of color distances, and displays the actual colors as
row and column annotations outside the color distance
heatmap.
}
\examples{
pc <- grDevices::palette.colors(25);
cd <- color_distance(pc, method="cie2000");
show_color_distance(cd, pc)

# with clustering
show_color_distance(cd, pc, cluster_data=TRUE,
   column_title_gp=grid::gpar(fontsize=20),
   column_title="palette.colors()")

# compare two color vectors
pc1 <- rainbowJam(10, preset="ryb2")
pc2 <- colorspace::rainbow_hcl(10)
cd <- color_distance(pc1, pc2)
show_color_distance(cd, cluster_data=TRUE,
   row_title_rot=90, row_title="rainbowJam()",
   row_title_gp=grid::gpar(fontsize=20),
   column_title_gp=grid::gpar(fontsize=20),
   column_title_rot=0, column_title="rainbow_hcl()")

# evaluate the small step size between HCL rainbow colors
show_color_distance(pc2,
   column_title_gp=grid::gpar(fontsize=20),
   column_title_rot=0, column_title="rainbow_hcl()")

# evaluate the larger step sizes between colorjam rainbow colors
show_color_distance(pc1, cluster_data=FALSE,
   column_title_gp=grid::gpar(fontsize=20),
   column_title_rot=0, column_title="rainbowJam()")

}
