% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-gradient.R
\name{showDichromat}
\alias{showDichromat}
\title{Show colors using dichromat color blindness adjustment}
\usage{
showDichromat(
  x,
  type = c("deutan", "protan", "tritan"),
  sep = "\\n",
  spacer = TRUE,
  original = TRUE,
  do_plot = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{list} or \code{character} vector with R compatible colors.}

\item{type}{\code{character} passed to \code{dichromat::dichromat()} for one
or more types of color blindness to simulate.}

\item{sep}{\code{character} used as a delimited to label each resulting
color vector.}

\item{spacer}{\code{logical} indicating whether to include a blank spacer
between sets of colors. This spacer is mainly useful for display.}

\item{original}{\code{logical} indicating whether to include original colors
and adjusted colors.}

\item{do_plot}{\code{logical} indicating whether to plot the results
using \code{jamba::showColors()}.}

\item{...}{additional arguments are ignored.}
}
\description{
Show colors using dichromat color blindness adjustment
}
\details{
This function is a very simple wrapper around \code{jamba::showColors()}
which also applies one of the color blindness emulations from
\code{dichromat::dichromat()}.
}
\examples{
showDichromat(jam_linear["firebrick"])

showDichromat(jam_linear[1:2])

showDichromat(jam_linear[7:9])

showDichromat(jam_linear, type="tritan", spacer=FALSE)

showDichromat(jam_linear, type="tritan", spacer=FALSE, original=FALSE)

}
\seealso{
Other colorjam display: 
\code{\link{color_pie}()}
}
\concept{colorjam display}
