% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-slot-colors.R
\name{slot_colors}
\alias{slot_colors}
\title{Slot a set of colors into a reference set of colors}
\usage{
slot_colors(
  given_colors,
  ref_colors,
  dist_threshold = NULL,
  min_distance = 11.5,
  return_type = c("list", "vector"),
  use_white = "F5",
  method = "cie2000",
  do_plot = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{given_colors}{\code{character} vector of given colors to test versus
\code{ref_colors}. Colors in \code{given_colors} which are within the distance
threshold of one or more colors in \code{ref_colors} are assigned (slotted)
into those colors.}

\item{ref_colors}{\code{character} vector of reference colors.}

\item{dist_threshold}{\code{numeric} default NULL, with pre-defined color
distance threshold.}

\item{min_distance}{\code{numeric} default 11.5 with a minimum distance to use,
when the dynamic distance threshold is below this value, and
when \code{dist_threshold} is not provided.}

\item{return_type}{\code{character} string, default 'list' with return type:
\itemize{
\item 'list': returns \code{list} with length(given_colors) and \code{integer} values
of assigned colors in \code{ref_colors}, or \code{NA} when no values in
\code{ref_colors} are within the distance threshold.
\item \code{vector}: returns vector equivalent of the \code{list} output,
flattened using \code{jamba::cPaste()}. Values are \code{NA} or 'NA' for
unslotted colors.
}}

\item{use_white}{\code{character} default "F5" representing the
white reference, any value recognized by \code{farver::as_white_ref()}.
\itemize{
\item The default 'F5' represents 'daylight fluorescent' and in
qualitative testing was most effective when defining color
distances.
\item The typical default 'D65' is 'daylight 6500K' and
is typically used for neutral daylight without blue (cool) or
yellow (warm) shifted background lighting.
}}

\item{method}{\code{character} default 'cie2000' passed to \code{color_distance()}}

\item{do_plot}{\code{logical} default FALSE, whether to plot a visual
of the slot colors assignments.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to internal functions
\code{color_distance()}, or when \code{do_plot==TRUE}, \code{imageByColors()}.}
}
\value{
\code{list} with length(given_colors) with \code{integer} vectors
referencing one or more elements in \code{ref_colors}, or \code{NULL}.
}
\description{
Slot a set of colors into a reference set of colors
}
\details{
The primary purpose is for \code{add_colors()} to determine whether
a given set of colors \code{given_colors} are already represented,
within a given color tolerance or distance,
within a reference set of colors \code{ref_colors}.

Colors are "slotted" into the corresponding reference colors,
and un-slotted colors can be considered "not used" in the reference
colors, and therefore they are available to be new colors
to add to the set.

\code{add_colors()} is recommended for most users, since it applies the
logic by calling \code{slot_colors()} itself. However, for debugging,
in order to understand why certain colors are "slotted" and other
colors are not, \code{slot_colors()} may be useful to call directly,
especially with \code{do_plot=TRUE} for visual review.

Color distance metrics are imperfect, and do not represent
color blindness conditions effectively. Further, color distance
metrics are designed with different goals in mind. For example
the typical metrics are designed to ensure tolerance to a
standard color. They are not primarily designed to quantify
the magnitude of perceptible difference between two colors,
and this latter scenario is the primary motivation of
\code{slot_colors()} and \code{add_colors()}.
}
\examples{
given_colors <- rainbowJam(5)
ref_colors <- colorspace::rainbow_hcl(6)
slot_colors(given_colors, ref_colors, do_plot=TRUE)

given_colors <- rainbowJam(5)
ref_colors <- colorspace::rainbow_hcl(16)
slot_colors(given_colors, ref_colors, verbose=TRUE, do_plot=TRUE)

}
