% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-presets.R
\name{validate_colorjam_preset}
\alias{validate_colorjam_preset}
\title{Validate h1,h2 color hue warp data}
\usage{
validate_colorjam_preset(
  h1 = NULL,
  h2 = NULL,
  direction = NULL,
  default_step = NULL,
  preset = NULL,
  digits = 4,
  ...
)
}
\arguments{
\item{h1, h2}{\code{numeric} color hue in degrees}

\item{direction}{\code{numeric} with one of two accepted values:
\itemize{
\item \code{1}: (default) indicating forward (increasing, clockwise)
progression around the color wheel
\item \code{-1}: indicating reverse
(decreasing, counter-clockwise) progression around the color
wheel.
}}

\item{digits}{\code{numeric} passed to \code{round()} before comparing
\code{h2} values for duplicated values.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} with components:
\itemize{
\item \code{h1}: HCL color hue indicating the actual hue to be used
when generating a color.
\item \code{h2}: HCL color hue indicating the virtual hue assigned to each
actual hue in \code{h1}.
\item \code{direction}: with \code{1} indicating forward (increasing, clockwise)
progression around the color wheel, and \code{-1} indicating reverse
(decreasing, counter-clockwise) progression around the color
wheel.
\item \code{default_step}: \code{character} string with name matching a
value returned by \code{colorjam_steps()}
}
}
\description{
Validate h1,h2 color hue warp data, internal function
}
\details{
Validate the \code{h1},\code{h2} color hue values alongside the \code{direction}
the angles should proceed. This function handles cases where
\code{h2} contains duplicate values, which is used to remove a
section of \code{h1} from the output \code{h2} color wheel. Duplicated
values are not handled well by \code{approx()}, so this function
rounds values to 4 digits, then sorts data based upon \code{direction},
then finally adjusts any duplicated \code{h2} values by adding
\code{1e-8}.
\subsection{Processing steps:}{
\itemize{
\item \code{h1},\code{h2} angles are adjusted within range \code{c(0, 360)} using
\code{x \%\% 360}.
\item \code{h1},\code{h2} values are rounded using \code{round(x, digits=digits)}
\item \code{h1},\code{h2} values are sorted by \code{h2} increasing, and \code{h1}
in order of \code{direction}
\item Duplicated \code{h2} values are adjusted by adding \code{c(0, 1, 2, ...) + 1e-8}
to each duplicated value, per set of duplicated values.
The first duplicated value in each set is unchanged, and
subsequent values in the set are increased by \code{1e-8}.
}
}

\subsection{TODO:}{
\itemize{
\item Enforce sensible ordering of values, mainly so the R-shiny app
h1,h2 edited values will permit pushing identical values.
For example, two identical \code{h2} values cause the corresponding
\code{h1} values to be skipped, thus removing that range of hues.
When one \code{h2} value is increased, it should also increase subsequent
values.
}
}
}
\seealso{
Other colorjam hue warp: 
\code{\link{add_colorjam_preset}()},
\code{\link{add_colorjam_step}()},
\code{\link{adjust_hue_warp}()},
\code{\link{approx_degrees}()},
\code{\link{colorjam_presets}()},
\code{\link{colorjam_steps}()},
\code{\link{display_degrees}()},
\code{\link{h2hwOptions}()},
\code{\link{h2hw}()},
\code{\link{hw2h}()},
\code{\link{mean_angle}()},
\code{\link{plot_colorjam_preset}()}
}
\concept{colorjam hue warp}
