% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-color.R
\name{vals2colorLevels}
\alias{vals2colorLevels}
\title{Apply color gradient to numeric values}
\usage{
vals2colorLevels(
  x,
  divergent = TRUE,
  col = "RdBu_r",
  defaultBaseColor = "#FFFFFF",
  lens = 0,
  numLimit = NULL,
  baseline = NULL,
  rampN = 25,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{numeric vector}

\item{divergent}{logical indicating whether the numeric values
are divergent, by default baseline=0 will center the color
ramp at zero.}

\item{col}{color value compatible with the \code{col} argument of
\code{jamba::getColorRamp()}. Example include: single color; multiple
colors; single color ramp name; or a custom color function.}

\item{defaultBaseColor}{character color used as a base color when
a single color is supplied in \code{col}.}

\item{lens}{numeric value sent to \code{jamba::warpRamp()}, to define the
level of color warping to apply to the color gradient, where \code{lens=0}
applies no adjustment.}

\item{numLimit}{numeric value indicating the maximum numeric value,
where values in \code{x} greater than this value are assigned to the
maximum color. When not defined, and \code{divergent=TRUE} it uses
\code{max(abs(x), na.rm=TRUE)}, or \code{divergent=FALSE} it uses
\code{max(x, na.rm=TRUE)}.}

\item{baseline}{numeric value indicating the minimum numeric value,
where values in \code{x} less than this value are assigned to the
minimum color. When not defined, and \code{divergent=TRUE} it sets
\code{baseline=0}; when \code{divergent=FALSE} it uses \code{min(x, na.rm=TRUE)}.}

\item{rampN}{integer number of colors to define for the color
gradient. Higher values define a smooth color gradient.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{jamba::getColorRamp()}.}
}
\description{
Apply color gradient to numeric values
}
\details{
This function is similar to several other existing R functions
that take a vector of numeric values, and apply a color gradient
(color ramp) to the numeric values. This function provides the ability
to warp the color ramp, for example using \code{jamba::warpRamp()} in order
to adjust the color gradient relative to the numeric range of the
data.
}
\examples{
# Start with an example numeric vector
x <- nameVector(-5:10);
showColors(vals2colorLevels(x));

# decrease the number of gradient colors
showColors(vals2colorLevels(x, rampN=15))

# change the baseline
showColors(vals2colorLevels(x, baseline=-2));

# adjust the gradient using lens
par("mar"=c(5,5,4,2));
imageByColors(rbindList(lapply(nameVector(c(-5,-2,0,2,5)), function(lens){
   vals2colorLevels(x, rampN=25, lens=lens);
})));
title(ylab="color lens factor", xlab="numeric value",
   main="Effects of warping the color gradient");

}
\seealso{
Other jam color functions: 
\code{\link{closestRcolor}()},
\code{\link{group2colors}()},
\code{\link{jam_pal}()},
\code{\link{matrix2heatColors}()},
\code{\link{rainbowJam}()},
\code{\link{scale_color_jam}()},
\code{\link{scale_fill_jam}()},
\code{\link{theme_jam}()}
}
\concept{jam color functions}
