% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-presets.R
\name{vibrant_color_by_hue}
\alias{vibrant_color_by_hue}
\title{Get vibrant reference color for a given HCL color hue}
\usage{
vibrant_color_by_hue(h, C = 60, S = 100, L = 50)
}
\arguments{
\item{h}{\code{numeric} color hue in the range \code{c(0, 360)}.}

\item{C, S, L}{\code{numeric} values used during the color conversion:
\itemize{
\item \code{C} is the intermediate HCL color chroma
\item \code{S} is the HSL color saturation
\item \code{L} is the HSL lightness
}}

\item{...}{additional arguments are ignored.}
}
\value{
\code{character} vector of colors for each hue \code{h}.
}
\description{
Get vibrant reference color for a given HCL color hue
}
\details{
This function converts each \code{h} HCL hue to a hex color using
\code{jamba::hcl2col()}. This hex color is converted to HSL color
space using \code{jamba::col2hsl()}, so the color saturation and
lightness can be adjusted to full saturation without
going outside the HCL color gamut. This HSL color is converted
back to hex color format.

The reason for the two-step conversion is to honor the HCL color
hue, which is not the same numeric value as used by the HSL
color space.
}
\seealso{
Other colorjam assignment: 
\code{\link{col_div_xf}()},
\code{\link{col_linear_xf}()},
\code{\link{group2colors}()},
\code{\link{matrix2heatColors}()},
\code{\link{rainbowJamMulti}()},
\code{\link{vals2colorLevels}()}
}
\concept{colorjam assignment}
