% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-presets.R
\name{add_colorjam_preset}
\alias{add_colorjam_preset}
\title{Add colorjam hue warp preset}
\usage{
add_colorjam_preset(
  preset,
  h1,
  h2 = NULL,
  direction = 1,
  default_step = NULL,
  description = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{preset}{\code{character} string with the preset name.}

\item{h1, h2}{\code{numeric} vectors of equal length, or \code{NULL} to
remove an existing preset.}

\item{direction}{\code{numeric} with one of two accepted values:
\itemize{
\item \code{1}: (default) indicating forward (increasing, clockwise)
progression around the color wheel
\item \code{-1}: indicating reverse
(decreasing, counter-clockwise) progression around the color
wheel.
}}

\item{default_step}{\code{character} string indicating the recommended
\code{step} from \code{colorjam_steps()}.}

\item{description}{\code{character} with optional text description of
the preset. The description will be added as an attribute:
\code{attr(x, "description")}}

\item{...}{additional arguments are ignored.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Add colorjam hue warp preset
}
\details{
The colorjam presets represent customized color wheels, used
for various color operations.
\itemize{
\item \code{h1} contains actual color hue values, with range \code{c(0, 360)}.
Its values are intended to be used in HCL color hue operations,
such as \code{jamba::hcl2col()}, or \code{colorspace::polarLUV()}. These
are not equal to \code{HSV} or \code{HLS} (also known as HSL) color hues,
although colorjam preset could be used to convert from HCL to HSV.
\item \code{h2} contains virtual color hue values, with range \code{c(0, 360)}.
These values are intended to represent the full virtual color
wheel, mapping the virtual hue \code{h2} to actual hue \code{h1}.
Not all values in \code{h1} need be represented, in fact \code{"dichromat"}
presets explicitly remove a sizable chunk of green color hues.
\item \code{direction} the direction of \code{h1} relative to \code{h2} encoded as:
\itemize{
\item \code{1} for forward direction, and
\item \code{-1} for reverse direction.
}
\item \code{default_step} used primarily by \code{rainbowJam()} to define the
appropriate Chroma/Luminance stepping pattern when choosing
actual HCL colors around the color wheel.
\item optional attribute \code{"description"} as \code{attr(x, "description")}.
}
}
\examples{
h1 <- c(8, 30, 65,
   120, 200, 240,
   260, 280, 330)
h2 <- c(115, 35.9, -3.7,
   -3.7, -3.7, -3.7,
   -70.9, -189.6, -229.2)
add_colorjam_preset("custom_dichromat",
   h1=h1, h2=h2, direction=-1, default_step="v24",
   description="Custom dichromat example")

# behold the new preset name appears
colorjam_presets()

# the preset data is available
colorjam_presets("custom_dichromat")

# the preset is used to create rainbow categorical colors
color_pie(rainbowJam(n=10,
   preset="custom_dichromat"))

# remove a preset by defining `h1=NULL`
add_colorjam_preset("custom_dichromat", h1=NULL)
colorjam_presets()

}
\seealso{
Other colorjam hue warp: 
\code{\link{add_colorjam_step}()},
\code{\link{adjust_hue_warp}()},
\code{\link{approx_degrees}()},
\code{\link{colorjam_presets}()},
\code{\link{colorjam_steps}()},
\code{\link{display_degrees}()},
\code{\link{h2hw}()},
\code{\link{h2hwOptions}()},
\code{\link{hcl_to_hsl_hue}()},
\code{\link{hsl_to_hcl_hue}()},
\code{\link{hw2h}()},
\code{\link{mean_angle}()},
\code{\link{plot_colorjam_preset}()},
\code{\link{plot_colorjam_steps}()},
\code{\link{remap_colorjam_preset}()},
\code{\link{validate_colorjam_preset}()}
}
\concept{colorjam hue warp}
