% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-color.R
\name{closestRcolor}
\alias{closestRcolor}
\title{Find the closest R color}
\usage{
closestRcolor(
  x,
  colorSet = unvigrep("^gr[ae]y($|[0-9]+$)", colors()),
  showPalette = FALSE,
  colorModel = c("hcl", "LUV"),
  Hwt = 2,
  Cwt = 1,
  Lwt = 4,
  warpHue = TRUE,
  method = "maximum",
  returnType = c("color", "name", "match"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{character vector of colors, either in hex format or any
valid color in R.}

\item{colorSet}{vector of colors, by default includes the R colors
provided by \code{grDevices::colors()} minus the "grey##" greyscale colors,
however any vector of colors can be used, whether named or not.}

\item{showPalette}{boolean indicating whether to display the input
colors and resulting closest matching colors by using
\code{jamba::showColors()}.}

\item{colorModel}{character color model to use, currently only using
"hcl", though "rgb" and "ryb" were both tested and judged to perform
less effectively.}

\item{Hwt, Cwt, Lwt}{relative weights for each dimension of HCL colors,
respectively.}

\item{warpHue}{boolean indicating whether to perform the hue warp
operation using \code{h2hw()} which improves the ability to match
colors between orange and green.}

\item{returnType}{character type of data to return: "color" will return
the actual closest color; "name" will return the name of the closest
color, or if the \code{colorSet} vector has no names, its values will
be used; "match" will return an integer vector as an index to colors
in \code{colorSet}.}

\item{verbose}{logical whether to print verbose output}
}
\description{
Find the closest R color for a vector of colors
}
\details{
This function is intended as a relatively efficient method to compare
a set of colors to the named R colors provided by \code{colors()}.
It has some distinction from similar methods, in that it uses HCL
color space, unlike similar methods which may use RGB color space.
It also prioritizes the color hue, then the luminance (visual brightness),
then chroma (saturation). These weights can be adjusted if necessary
although the default values appear to work fairly well.
}
\examples{
closestRcolor(rainbowJam(12), showPalette=TRUE);

}
\seealso{
Other jam color functions: 
\code{\link{group2colors}()},
\code{\link{jam_pal}()},
\code{\link{matrix2heatColors}()},
\code{\link{rainbowJam}()},
\code{\link{scale_color_jam}()},
\code{\link{scale_fill_jam}()},
\code{\link{theme_jam}()},
\code{\link{vals2colorLevels}()}
}
\concept{jam color functions}
