% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-color-distance.R
\name{color_distance}
\alias{color_distance}
\title{Calculate color distance between two or more colors}
\usage{
color_distance(
  x,
  y = NULL,
  method = c("cie2000", "cie94", "cie1976", "cmc", "euclidean"),
  use_white = "F5",
  do_plot = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{character} color, required input colors.
\itemize{
\item When \code{y} is supplied, values in \code{y} are recycled to \code{length(x)},
and each entry in \code{x} is directly compared to \code{y}.
\item When \code{y} is not supplied, \code{x} is compared to itself,
returning a \code{matrix}.
}}

\item{y}{\code{character} color, default NULL}

\item{method}{\code{character}, default 'cie2000', the color distance method,
passed to \code{farver::compare_colour()}. In future, this argument
may permit additional distance methods, and/or specific color function,
in order to impose other criteria and adjustments.}

\item{use_white}{\code{character} default "F5" representing the
white reference, any value recognized by \code{farver::as_white_ref()}.
\itemize{
\item The default 'F5' represents 'daylight fluorescent' and in
qualitative testing was most effective when defining color
distances.
\item The typical default 'D65' is 'daylight 6500K' and
is typically used for neutral daylight without blue (cool) or
yellow (warm) shifted background lighting.
}}

\item{...}{additional arguments are passed to \code{farver::compare_colour()}}
}
\value{
\code{numeric} color distance with \code{length(x)} entries when both
\code{x} and \code{y} are supplied, or a \code{numeric} matrix with color distances
between all entries in \code{x}.
\itemize{
\item An attribute 'method' is added with the color distance used,
mainly so the method can be used by \code{show_color_distance()},
but also by other methods as relevant.
}
}
\description{
Calculate color distance between two or more colors
}
\details{
Color distance is calculated using \code{farver::compare_colour()},
with some defaults intended in future to assist with color blindness
calculations.
}
\examples{
color_distance("red", "firebrick", use_white="D65")

color_distance(grDevices::palette.colors(15))

pc <- rainbowJam(5)
cd <- color_distance(pc);
show_color_distance(cd, pc=pc, column_title="rainbowJam()");

pc <- grDevices::palette.colors(15);
cd <- color_distance(pc);
show_color_distance(cd, pc=pc, column_title="palette.colors()");

}
