% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-color-distance.R
\name{color_distance}
\alias{color_distance}
\title{Calculate color distance between two or more colors}
\usage{
color_distance(
  x,
  y = NULL,
  method = c("cmc", "cie2000", "cie94", "cie1976", "euclidean"),
  use_white = "F5",
  lightness = 2/3,
  chroma = 1,
  do_plot = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{character} color, required input colors.
\itemize{
\item When \code{y} is supplied, values in \code{y} are recycled to \code{length(x)},
and each entry in \code{x} is directly compared to \code{y}.
\item When \code{y} is not supplied, \code{x} is compared to itself,
returning a \code{matrix}.
}}

\item{y}{\code{character} color, default NULL}

\item{method}{\code{character}, default 'cmc', the color distance method,
passed to \code{farver::compare_colour()}.
\itemize{
\item Note that 'cmc' uses arguments \code{lightness} and \code{chroma} which
have custom values 2/3 and 1, respectively. The defaults are 2 and 1
for acceptability, and 1 and 1 for perceptability.
In our testing, 2/3 and 1 were more useful, since lightness on
computer screen and in print is more easily perceived than
on a painted surface.
\item In future, this argument
may permit additional distance methods, and/or specific color function,
in order to impose other criteria and adjustments.
}}

\item{use_white}{\code{character} default "F5" representing the
white reference, any value recognized by \code{farver::as_white_ref()}.
\itemize{
\item The default 'F5' represents 'daylight fluorescent' and in
qualitative testing was most effective when defining color
distances.
\item The typical default 'D65' is 'daylight 6500K' and
is typically used for neutral daylight without blue (cool) or
yellow (warm) shifted background lighting.
}}

\item{do_plot}{\code{logical} default FALSE, whether to plot results using
\code{show_color_distance()}.}

\item{...}{additional arguments are passed to \code{farver::compare_colour()}
and to \code{show_color_distance()}.}
}
\value{
\code{numeric} color distance with \code{length(x)} entries when both
\code{x} and \code{y} are supplied, or a \code{numeric} matrix with color distances
between all entries in \code{x}.
\itemize{
\item An attribute 'method' is added with the color distance used,
mainly so the method can be used by \code{show_color_distance()},
but also by other methods as relevant.
}
}
\description{
Calculate color distance between two or more colors
}
\details{
Color distance is calculated using \code{farver::compare_colour()},
with some defaults intended in future to assist with color blindness
calculations.
}
\examples{
color_distance("red", "firebrick", use_white="D65")

palette15 <- sort_colors(grDevices::palette.colors(15, "Polychrome 36"))
color_distance(palette15, do_plot=TRUE)

pc <- rainbowJam(8)
cd <- color_distance(pc);
show_color_distance(cd, pc=pc);

}
\seealso{
Other colorjam internal: 
\code{\link{find_color_spread}()},
\code{\link{jam_pal}()},
\code{\link{rainbowJam_v1}()},
\code{\link{show_color_distance}()},
\code{\link{slot_colors}()},
\code{\link{vals2colorLevels}()}
}
\concept{colorjam internal}
