% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-rainbowjam.R
\name{color_pie}
\alias{color_pie}
\title{Show colors spread around a pie chart}
\usage{
color_pie(
  colors,
  border = colors,
  lwd = 2,
  radius = 1.5,
  label_radius = radius * 0.75,
  add = FALSE,
  ...
)
}
\arguments{
\item{colors}{\code{vector} of R colors.}

\item{border}{\code{vector} of R colors used to draw a border around
each pie wedge. By default it uses input \code{colors}.}

\item{lwd}{numeric value used to define the line width of the
pie wedge borders.}

\item{radius}{numeric value representing the radius of the
overall pie chart, where \code{radius=1} represents the default
radius used by \code{graphics::pie()}. The default is \code{radius=1.5}
in order to use more of the output plot size.}

\item{label_radius}{numeric value indicating the radius used
for labels, intended to allow labels to appear inside each
pie wedge.}

\item{add}{logical indicating whether to draw the pie chart
onto the existing plot device, without creating a new plot.}

\item{...}{additional arguments are passed to \code{graphics::pie()}.}
}
\description{
Show colors spread around a pie chart
}
\details{
This function simply displays colors in a pie chart
format.

If the input is a \code{list}, each list is used to produce
layers of a pie chart, in order to help compare colors
from each vector in the list.
}
\examples{
color_pie(rainbowJam(20, nameStyle="none"),
   sub="rainbowJam(20)")

n <- 8;
color_pie(rainbowJam(n),
   sub="rainbowJam(8)")

color_pie(rainbow(n),
   sub="rainbow(8)")

color_pie(colorspace::rainbow_hcl(n),
   sub="colorspace::rainbow_hcl(8)")

color_pie(colorspace::rainbow_hcl(n, c=120),
   sub="colorspace::rainbow_hcl(8, c=120)")

color_list <- list(rainbowJam=rainbowJam(n),
   rainbow_hcl=colorspace::rainbow_hcl(n, c=120));
color_pie(color_list,
   sub="inside ring:rainbow_hcl()\nouter ring: rainbowJam()")

rainbow_list <- lapply(4*c(5,4,2,1), function(n){
   rainbowJam(n, nameStyle="none");
});
color_pie(rainbow_list,
   sub="rainbowJam()\nn=4, 8, 16, 20")

}
\seealso{
Other jam color functions: 
\code{\link{closestRcolor}()},
\code{\link{group2colors}()},
\code{\link{jam_pal}()},
\code{\link{matrix2heatColors}()},
\code{\link{rainbowJam_v1}()},
\code{\link{rainbowJam}()},
\code{\link{scale_color_jam}()},
\code{\link{scale_fill_jam}()},
\code{\link{theme_jam}()},
\code{\link{vals2colorLevels}()}
}
\concept{jam color functions}
