% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-steps.R
\name{colorjam_steps}
\alias{colorjam_steps}
\title{Colorjam chroma/luminance steps}
\usage{
colorjam_steps(step = NULL, ...)
}
\arguments{
\item{...}{additional arguments are ignored.}

\item{steps}{\code{NULL} to return a \code{character} vector of all recognized
steps, or \code{character} string to return specific data associated
with a recognized steps name.}
}
\value{
\code{character} vector of recognized colorjam step names when
\code{step} is \code{NULL}, or
when \code{step} is defined a \code{list} is returned with elements
\itemize{
\item \code{C}: \code{numeric} values indicating HCL Chroma
\item \code{L}: \code{numeric} values indicating HCL Chroma
}
}
\description{
Colorjam chroma/luminance steps to adjust a series of color hues
to visibly distinct categorical colors.
}
\details{
\code{colorjam_steps()}: list the names of available colorjam steps,
or when a \code{step} name is provided, it returns a \code{list} with elements
\code{"C"} and \code{"L"}.
}
\seealso{
Other colorjam hue warp: 
\code{\link{add_colorjam_preset}()},
\code{\link{add_colorjam_step}()},
\code{\link{adjust_hue_warp}()},
\code{\link{approx_degrees}()},
\code{\link{colorjam_presets}()},
\code{\link{display_degrees}()},
\code{\link{h2hw}()},
\code{\link{h2hwOptions}()},
\code{\link{hcl_to_hsl_hue}()},
\code{\link{hsl_to_hcl_hue}()},
\code{\link{hw2h}()},
\code{\link{mean_angle}()},
\code{\link{plot_colorjam_preset}()},
\code{\link{remap_colorjam_preset}()},
\code{\link{validate_colorjam_preset}()}
}
\concept{colorjam hue warp}
