% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-color.R
\name{h2hw}
\alias{h2hw}
\title{Convert standard hue to warped hue}
\usage{
h2hw(h, h1 = h2hwOptions()$h1, h2 = h2hwOptions()$h2, ...)
}
\arguments{
\item{h}{numeric vector of color hues between 0 and 360. These hues do
not need to be in sequential order.}

\item{h1, h2}{vector of color hues, which by default are defined in
\code{\link[=h2hwOptions]{h2hwOptions()}}, but allowed here in cases where the global options
should be overridden but not modified.}
}
\value{
numeric vector of hue values after applying the hue warp
operation.

numeric vector of warped color hues.
}
\description{
Convert standard hue to warped hue using the hue warp vectors
}
\details{
This function is intended to convert from a vector of hue values to
the warped hues defined by the vectors returned by \code{\link[=h2hwOptions]{h2hwOptions()}}.
The intent is to convert colors in RGB space into RYB space by default,
which substantially improves several other color manipulations, such
as selection of categorical colors, and color blending.

Note the input hue is considered the "standard" color hue as defined
by the \code{\link[colorspace:polarLUV]{colorspace::polarLUV()}} function, ranging between 0 and 360.
By this standard, 0 is defined as red, 120 is defined as green, and
240 is defined as blue.

The default mappings convert RGB (red, green, blue) to RYB (red, yellow,
blue). The color yellow has hue=60 in RGB space, so the call to
\code{hw2h(60)} results in \code{120}, which is the hue in RYB space.
}
\examples{
## Yellow when using an RGB color wheel is 60 degrees,
## but on an RYB color wheel is 120 degrees.
h2hw(60);

# RGB colors are convenient, but are not ideal especially when blending
# colors. Note that blue and yellow have hues that differ by exactly 180
# degrees, meaning a hue average is as likely to be purple as green.
huesBY <- col2hcl(c("blue", "yellow"))["H",];
huesBY;
warpedHuesBY <- h2hw(huesBY);
warpedHuesBY;
warpedHues <- h2hw(hues);

}
\seealso{
Other hue warp functions: \code{\link{h2hwOptions}},
  \code{\link{hw2h}}
}
\concept{hue warp functions}
