% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-huewarp.R
\name{hw2h}
\alias{hw2h}
\title{Convert warped hue to standard hue}
\usage{
hw2h(h, h1 = h2hwOptions()$h1, h2 = h2hwOptions()$h2, ...)
}
\arguments{
\item{h}{numeric vector of color hues between 0 and 360. These hues do
not need to be in sequential order.}

\item{h1, h2}{vector of color hues, which by default are defined in
\code{\link[=h2hwOptions]{h2hwOptions()}}, but allowed here in cases where the global options
should be overridden but not modified.}
}
\value{
numeric vector of color hues.
}
\description{
Convert warped hue to standard hue using the hue warp vectors
}
\details{
This function is intended to convert from a vector of warped hue values to
the hues defined by the vectors returned by \code{h2hwOptions}.
The intent is to convert colors in RYB space into RGB space by default.

One example of input would be to supply a uniformly spaced set of color
hues, and convert them to "standard" hues by HCL standards. It has the
effect of presenting categorical colors, using a non-linear set of
"standard" hue values.

The default mappings convert RYB (red, yellow, blue) to RGB (red, green,
blue). The color yellow has hue=120 in RYB space, so the call to
\code{h2hw(120)} results in 60, which is the hue in RGB space.
}
\examples{
# It can be useful to create a uniform sequence of angles in warped
# hues, which are visually more uniform than those using an RGB color wheel,
# then convert those hues to standard color hues.
warpedHues <- seq(from=0, to=330, length.out=12);
warpedHues;
hues <- hw2h(warpedHues);
hues;

}
\seealso{
Other hue warp functions: 
\code{\link{h2hwOptions}()},
\code{\link{h2hw}()}
}
\concept{hue warp functions}
