% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-color.R
\name{jam_pal}
\alias{jam_pal}
\title{Jam color palette for ggplot2}
\usage{
jam_pal(type = "seq", palette = 1, direction = 1, invert = FALSE,
  useGrey = 20)
}
\arguments{
\item{type}{character string indicating the colors are sequential
\code{"seq"}.}

\item{palette}{integer value indicating the categorical palette
to use, intended to provide variety in the color assignment.
(Not yet implemented.)}

\item{direction}{integer indicating whether to reverse the color
assignment, either \code{1} for the default forward assignment, or
\code{-1} for reverse assignment. Any negative value will reverse
the colors.}

\item{invert}{logical indicating whether to return corresponding
contrasting colors, for example for text labels, typically either
\code{"white"} or \code{"black"} as defined by \code{jamba::setTextContrastColor()}.}

\item{useGrey}{integer value between 0 and 100 indicating the grey
value, as sent to \code{jamba::setTextContrastColor()}, used only when
\code{invert=TRUE}.}
}
\description{
Jam color palette for ggplot2
}
