% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-ggplot2-themejam.R
\name{scale_color_jam}
\alias{scale_color_jam}
\title{Apply rainbowJam categorical colors to a ggplot2 object}
\usage{
scale_color_jam(
  ...,
  type = "seq",
  palette = 1,
  direction = 1,
  invert = FALSE,
  darkFactor = 1,
  sFactor = 1,
  darken = FALSE,
  alpha = 1,
  useGrey = 20,
  preset = getOption("colorjam.preset", "dichromat2"),
  step = getOption("colorjam.step", "default")
)
}
\arguments{
\item{...}{additional arguments are passed to \code{ggplot2::discrete_scale()}.}

\item{type}{\code{character} string indicating the colors are sequential
\code{"seq"}, and is passed to \code{colorjam::jam_pal()}.}

\item{palette}{\code{integer} value indicating the categorical palette
to use, intended to provide variety in the color assignment.
(Not yet implemented.)}

\item{direction}{\code{integer} indicating whether to reverse the color
assignment, either \code{1} for the default forward assignment, or
\code{-1} for reverse assignment. Any negative value will reverse
the colors.}

\item{invert}{\code{logical} indicating whether to return corresponding
contrasting colors, for example for text labels, typically either
\code{"white"} or \code{"black"} as defined by \code{jamba::setTextContrastColor()}.}

\item{darkFactor, sFactor}{\code{numeric} passed to \code{jamba::makeColorDarker()}
for optional adjustment of the color, by darkness and saturation,
respectively. Intended when using \code{scale_color_jam()} and
\code{scale_fill_jam()} where you want the color value to be lighter
or darker than the fill color, a useful effect for outlines.}

\item{darken}{\code{logical} indicating whether to apply default values
for \code{darkFactor} and \code{sFactor} to darken the resulting colors.}

\item{alpha}{\code{numeric} value indicating the alpha transparency, on a
scale of 0 (transparent) to 1 (non-transparent).}

\item{useGrey}{\code{integer} value between 0 and 100 indicating the grey
value, as sent to \code{jamba::setTextContrastColor()}, used only when
\code{invert=TRUE}.}
}
\description{
Apply rainbowJam categorical colors to a ggplot2 object
}
\details{
This function provides a function in the format \verb{scale_color_*}
to be applied to ggplot2 objects. It can provide a more visibly
distinct set of categorical colors than \code{ggplot2::scale_color_hue()}.
}
\examples{
if (jamba::check_pkg_installed("ggplot2")) {
   dsamp <- ggplot2::diamonds[sample(nrow(ggplot2::diamonds), 1000),];
   d <- ggplot2::ggplot(
      dsamp, ggplot2::aes(carat, price)) +
      ggplot2::geom_point(ggplot2::aes(colour=cut),
         size=4);

   print(d + ggplot2::scale_color_hue() + ggplot2::ggtitle("scale_color_hue()"));
   print(d + scale_color_jam() + ggplot2::ggtitle("scale_color_jam()"));
   print(d + scale_color_jam(preset="ryb") + ggplot2::ggtitle("scale_color_jam(preset='ryb')"));
}

}
\seealso{
Other colorjam ggplot2: 
\code{\link{jam_pal}()},
\code{\link{scale_fill_jam}()},
\code{\link{theme_jam}()}
}
\concept{colorjam ggplot2}
