% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-add-colors.R
\name{show_color_distance}
\alias{show_color_distance}
\title{Show color distance as a heatmap}
\usage{
show_color_distance(
  cd = NULL,
  pc = NULL,
  show_labels = TRUE,
  cluster_data = FALSE,
  row_split = 0,
  ...
)
}
\description{
Show color distance as a heatmap
}
\examples{
pc <- grDevices::palette.colors(25);
cd <- color_distance(pc, method="cmc");
cd <- color_distance(pc, method="euclidean");
cd <- color_distance(pc, method="cie2000");
show_color_distance(cd, pc, cluster_data=TRUE, row_split=4)

show_color_distance(cd, pc, cluster_data=TRUE, row_split=5,
clustering_distance_rows="euclidean", clustering_distance_columns="euclidean")
cd2 <- 360 - cd;
diag(cd2) <- 360;
show_color_distance(cd2, pc, cluster_data=TRUE, row_split=4,
clustering_distance_rows="euclidean", clustering_distance_columns="euclidean")

show_color_distance(cd, pc)

pc <- rainbowJam(5)
}
