% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-show-color-distance.R
\name{show_color_distance}
\alias{show_color_distance}
\title{Show color distance as a heatmap}
\usage{
show_color_distance(
  cd = NULL,
  pc = NULL,
  show_labels = TRUE,
  cluster_data = FALSE,
  row_split = 0,
  ...
)
}
\arguments{
\item{cd}{\code{numeric} matrix, default NULL, with pre-calculated values.}

\item{pc}{\code{character} vector of colors used in \code{cd}, useful when
there are \code{names(pc)} to use as heatmap column or row labels.
When only \code{pc} is provided and not \code{cd}, the \code{cd} is calculated
by calling \code{color_distance(pc)}.}

\item{show_labels}{\code{logical} default TRUE, used when \code{pc} is provided.}

\item{cluster_data}{\code{logical} default FALSE, whether to enable row
and column hierarchical clustering in the heatmap.}

\item{row_split}{\code{numeric} default 0, used when \code{cluster_data=TRUE}
to subdivide the dendrogram into this many separate subclusters.}

\item{...}{additional arguments are passed to \code{color_distance()},
then to \code{jamses::heatmap_se()}. Many arguments in \code{jamses::heatmap_se()}
are also passed through to \code{ComplexHeatmap::Heatmap()}.}
}
\value{
\code{list} with two elements, suitable to use with
\code{jamba::imageByColors()}.
\itemize{
\item \code{cd}: the color distance matrix
\item \code{cdc}: the heatmap colorized matrix
}
}
\description{
Show color distance as a heatmap
}
\details{
This function formerly used \code{ComplexHeatmap::Heatmap()}, now uses
\code{jamba::imageByColors()} to remove the dependency on another
R package.

It creates a matrix visual
summary of color distances, and displays the actual colors as
row and column annotations outside the color distance
heatmap.
}
\examples{
\dontshow{if ((requireNamespace("ComplexHeatmap", quietly=TRUE))) withAutoprint(\{ # examplesIf}
pc <- grDevices::palette.colors(25, palette="Polychrome 36");
cd <- color_distance(pc, method="cie2000");
show_color_distance(cd, pc)

# with clustering
show_color_distance(cd, pc, cluster_data=TRUE)

# compare two color vectors
pc1 <- rainbowJam(10, preset="ryb2")
pc2 <- colorspace::rainbow_hcl(10)
cd <- color_distance(pc1, pc2)
show_color_distance(cd, cluster_data=FALSE)

# evaluate the small step size between HCL rainbow colors
show_color_distance(pc2)

# evaluate the larger step sizes between colorjam rainbow colors
show_color_distance(pc1, cluster_data=FALSE)
\dontshow{\}) # examplesIf}
}
\seealso{
Other colorjam internal: 
\code{\link{color_distance}()},
\code{\link{find_color_spread}()},
\code{\link{jam_pal}()},
\code{\link{rainbowJam_v1}()},
\code{\link{slot_colors}()},
\code{\link{vals2colorLevels}()}
}
\concept{colorjam internal}
