% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorjam-add-colors.R
\name{slot_colors}
\alias{slot_colors}
\title{Slot a set of colors into a reference set of colors}
\usage{
slot_colors(
  given_colors,
  ref_colors,
  dist_threshold = NULL,
  verbose = FALSE,
  ...
)
}
\value{
\code{list} with length(given_colors) with \code{integer} vectors
referencing one or more elements in \code{ref_colors}, or \code{NULL}.
}
\description{
Slot a set of colors into a reference set of colors
}
\examples{
given_colors <- rainbowJam(5)
ref_colors <- colorspace::rainbow_hcl(6)
slot_colors(given_colors, ref_colors)

}
