% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-color.R
\name{theme_jam}
\alias{theme_jam}
\title{Jam default theme for ggplot2}
\usage{
theme_jam(
  theme_default = ggplot2::theme_bw,
  base_size = 18,
  grid.major.size = 0.5,
  grid.minor.size = 0.25,
  strip.background.colour = "grey30",
  strip.background.fill = "lightgoldenrod1",
  panel.grid.major.colour = "grey80",
  panel.grid.minor.colour = "grey90",
  axis.text.x.angle = 60,
  blankGrid = FALSE,
  blankXgrid = FALSE,
  blankYgrid = FALSE,
  resetTheme = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{theme_default}{function representing a ggplot2 theme.}

\item{base_size}{default font point size, used for scaling the
overall text sizes larger or smaller.}

\item{grid.major.size, grid.minor.size}{the line width for the major
and minor grid lines, respectively. Set to 0 to suppress either.}

\item{strip.background.colour, strip.background.fill}{color for the
border and strip background itself when ggplot2 is using a
faceted layout.}

\item{panel.grid.major.colour, panel.grid.minor.colour}{colors for
the major and minor grid lines, respectively.}

\item{axis.text.x.angle}{numeric degrees to rotate the x-axis
labels, apparently starts at 0 (horizontal) and goes
counter-clockwise (to the left.)}

\item{blankGrid, blankXgrid, blankYgrid}{logical indicating whether
to have a blank grid for everything, major, or minor axis lines,
respectively. Intended to make it fast and easy to remove all
gridlines.}

\item{resetTheme}{logical whether to call the function \code{theme_default}
which essentially resets (replaces) all previous settings with
those defined in the theme function. If \code{FALSE} then only the
specific settings defined in this function will be applied.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{ggplot2::theme()} in
order to allow custom settings beyond what this function provides.}
}
\description{
Jam default theme for ggplot2
}
\details{
This function applies some default theme settings to ggplot2, mainly
taking away the default grey newspaper background color, also rotates the
x-axis label text to 60 degrees, to accomodate longer labels without
overlaps.
}
\examples{
if (suppressPackageStartupMessages(require(ggplot2))) {
   dsamp <- diamonds[sample(nrow(diamonds), 1000),];
   (d <- ggplot(dsamp, aes(carat, price)) + geom_point(aes(colour=cut), size=2));

   print(d + scale_color_jam() + ggtitle("scale_color_jam()"));
   print(d + scale_color_jam() + theme_jam() + ggtitle("scale_color_jam() + \ntheme_jam()"));
}

}
\seealso{
Other colorjam ggplot2: 
\code{\link{jam_pal}()},
\code{\link{scale_color_jam}()},
\code{\link{scale_fill_jam}()}
}
\concept{colorjam ggplot2}
